Crate iroh

Send data over the internet.

Getting started

Example

Create a new node and add some data to the blobs store. This data will be available over the network.

let node = iroh::node::Node::memory().spawn().await?;
let client = node.client();
let hash = client.blobs().add_bytes(b"some data".to_vec()).await?.hash;
println!("hash: {}", hash);

Explanation

Iroh node

To create an iroh Node, you use the Builder to configure the node and to spawn it.

There are also shortcuts to create an in memory or persistent node with default settings.

Iroh client

A node is controlled via a client. The client provides the main API to interact with a node, no matter if it is a local in-process node or a node in a different process. All clients are cheaply cloneable and can be shared across threads.

A handle to the client is available via the client method on the node.

Node also implements Deref to the client, so you can call client methods directly on the node.

Subsystems

The client provides access to various subsystems:

node: information and control of the iroh node itself

blobs: manage and share content-addressed blobs of data

tags: tags to tell iroh what data is important

gossip: exchange data with other nodes via a gossip protocol

authors: interact with document authors

docs: interact with documents

The subsystem clients can be obtained cheaply from the main iroh client. They are also cheaply cloneable and can be shared across threads.

So if you have code that only needs to interact with one subsystem, pass it just the subsystem client.

Remote nodes

To obtain a client to a remote node, you can use connect to connect to a node running on the same machine, using the given data directory, or connect_addr to connect to a node at a known address.

Important: the protocol to a remote node is not stable and will frequently change. So the client and server must be running the same version of iroh.

Reexports

The iroh crate re-exports the following crates:

iroh_base as base

iroh_blobs as blobs

iroh_docs as docs

iroh_gossip as gossip

iroh_net as net

Feature Flags

metrics: Enable metrics collection. Enabled by default.

fs-store: Enables the disk based storage backend for iroh-blobs. Enabled by default.

Modules

base

Base types and utilities for Iroh

blobs

Blobs layer for iroh.

client

Client to an Iroh node.

docs

Multi-dimensional key-value documents with an efficient synchronization protocol

gossip

Broadcast messages to peers subscribed to a topic

metricsmetrics

Expose metrics module

net

Peer-to-peer connectivity based on QUIC.

node

Node API

util

Utilities for working with iroh.