iroh::address_lookup

Trait IntoAddressLookup

pub trait IntoAddressLookup:
    Send
    + Sync
    + Debug
    + 'static {
    // Required method
    fn into_address_lookup(
        self,
        endpoint: &Endpoint,
    ) -> Result<impl AddressLookup, IntoAddressLookupError>;
}

Trait for structs that can be converted into AddressLookups.

This trait is implemented on builders for Address Lookup’s. Any type that implements this trait can be added as a Address Lookup in Builder::address_lookup.

Any type that implements AddressLookup also implements IntoAddressLookup.

Iroh uses this trait to allow configuring the set of address lookup services on the endpoint builder, while also providing them access to information about the endpoint to IntoAddressLookup::into_address_lookup.

Required Methods

fn into_address_lookup( self, endpoint: &Endpoint, ) -> Result<impl AddressLookup, IntoAddressLookupError>

Turns this AddressLookup builder into a ready-to-use Address Lookup.

If an error is returned, building the endpoint will fail with this error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.