Enum iroh::blobs::cli::BlobCommands

pub enum BlobCommands {
    Add {
        source: BlobSource,
        options: BlobAddOptions,
    },
    Get {
        ticket: TicketOrHash,
        address: Vec<SocketAddr>,
        relay_url: Option<RelayUrl>,
        recursive: Option<bool>,
        override_addresses: bool,
        node: Option<PublicKey>,
        out: Option<OutputTarget>,
        stable: bool,
        tag: Option<String>,
        queued: bool,
    },
    Export {
        hash: Hash,
        out: OutputTarget,
        recursive: bool,
        stable: bool,
    },
    List(ListCommands),
    Validate {
        verbose: u8,
        repair: bool,
    },
    ConsistencyCheck {
        verbose: u8,
        repair: bool,
    },
    Delete(DeleteCommands),
    Share {
        hash: Hash,
        addr_options: AddrInfoOptions,
        recursive: bool,
        debug: bool,
    },
}

Subcommands for the blob command.

Variants

Add

Add data from PATH to the running node.

Fields

source: BlobSource

Path to a file or folder.

If set to STDIN, the data will be read from stdin.

options: BlobAddOptions

Get

Download data to the running node’s database and provide it.

In addition to downloading the data, you can also specify an optional output directory where the data will be exported to after it has been downloaded.

Fields

ticket: TicketOrHash

Ticket or Hash to use.

address: Vec<SocketAddr>

Additional socket address to use to contact the node. Can be used multiple times.

relay_url: Option<RelayUrl>

Override the relay URL to use to contact the node.

recursive: Option<bool>

Override to treat the blob as a raw blob or a hash sequence.

override_addresses: bool

If set, the ticket’s direct addresses will not be used.

node: Option<PublicKey>

NodeId of the provider.

out: Option<OutputTarget>

Directory or file in which to save the file(s).

If set to STDOUT the output will be redirected to stdout.

If not specified, the data will only be stored internally.

stable: bool

If set, the data will be moved to the output directory, and iroh will assume that it will not change.

tag: Option<String>

Tag to tag the data with.

queued: bool

If set, will queue the download in the download queue.

Use this if you are doing many downloads in parallel and want to limit the number of downloads running concurrently.

Export

Export a blob from the internal blob store to the local filesystem.

Fields

hash: Hash

The hash to export.

out: OutputTarget

Directory or file in which to save the file(s).

If set to STDOUT the output will be redirected to stdout.

recursive: bool

Set to true if the hash refers to a collection and you want to export all children of the collection.

stable: bool

If set, the data will be moved to the output directory, and iroh will assume that it will not change.

List(ListCommands)

List available content on the node.

Validate

Validate hashes on the running node.

Fields

verbose: u8

Verbosity level.

repair: bool

Repair the store by removing invalid data

Caution: this will remove data to make the store consistent, even if the data might be salvageable. E.g. for an entry for which the outboard data is missing, the entry will be removed, even if the data is complete.

ConsistencyCheck

Perform a database consistency check on the running node.

Fields

verbose: u8

Verbosity level.

repair: bool

Repair the store by removing invalid data

Caution: this will remove data to make the store consistent, even if the data might be salvageable. E.g. for an entry for which the outboard data is missing, the entry will be removed, even if the data is complete.

Delete(DeleteCommands)

Delete content on the node.

Share

Get a ticket to share this blob.

Fields

hash: Hash

Hash of the blob to share.

addr_options: AddrInfoOptions

Options to configure the address information in the generated ticket.

Use relay-and-addresses in networks with no internet connectivity.

recursive: bool

If the blob is a collection, the requester will also fetch the listed blobs.

debug: bool

Display the contents of this ticket too.

Implementations

impl BlobCommands

pub async fn run(self, blobs: &Client, addr: NodeAddr) -> Result<(), Error>

Runs the blob command given the iroh client.