Function iroh::blobs::store::bao_tree::io::full_chunk_groups

pub fn full_chunk_groups(
    ranges: &RangeSet<[ChunkNum; 2]>,
    block_size: BlockSize
) -> RangeSet<[ChunkNum; 2]>

Given a range set of byte ranges, round it up to chunk groups.

If we store outboard data at a level of granularity of block_size, we can only share full chunk groups because we don’t have proofs for anything below a chunk group.