Trait iroh::blobs::store::bao_tree::io::fsm::CreateOutboard

pub trait CreateOutboard {
    // Required methods
    fn create_sized(
        data: impl AsyncStreamReader,
        size: u64,
        block_size: BlockSize
    ) -> impl Future<Output = Result<Self, Error>>
       where Self: Sized + Default;
    fn init_from(
        &mut self,
        data: impl AsyncStreamReader
    ) -> impl Future<Output = Result<(), Error>>;

    // Provided method
    async fn create(
        data: impl AsyncSliceReader,
        block_size: BlockSize
    ) -> Result<Self, Error>
       where Self: Sized + Default { ... }
}

Convenience trait to initialize an outboard from a data source.

In complex real applications, you might want to do this manually.

Required Methods

fn create_sized( data: impl AsyncStreamReader, size: u64, block_size: BlockSize ) -> impl Future<Output = Result<Self, Error>>
where Self: Sized + Default,

create an outboard from a data source. This requires the outboard to have a default implementation, which is the case for the memory implementations.

fn init_from( &mut self, data: impl AsyncStreamReader ) -> impl Future<Output = Result<(), Error>>

Init the outboard from a data source. This will use the existing tree and only init the data and set the root hash.

So this can be used to initialize an outboard that does not have a default, such as a file based one.

It will only include data up the the current tree size.

Provided Methods

async fn create( data: impl AsyncSliceReader, block_size: BlockSize ) -> Result<Self, Error>
where Self: Sized + Default,

Create an outboard from a seekable data source, measuring the size first.

This requires the outboard to have a default implementation, which is the case for the memory implementations.

Object Safety

This trait is not object safe.