Trait iroh::blobs::store::bao_tree::io::fsm::OutboardMut

pub trait OutboardMut: Sized {
    // Required methods
    fn save(
        &mut self,
        node: TreeNode,
        hash_pair: &(Hash, Hash)
    ) -> impl Future<Output = Result<(), Error>>;
    fn sync(&mut self) -> impl Future<Output = Result<(), Error>>;
}

A mutable outboard.

This trait provides a way to save a hash pair for a node and to set the length of the data file.

This trait can be used to incrementally save an outboard when receiving data. If you want to just ignore outboard data, there is a special placeholder outboard implementation super::outboard::EmptyOutboard.

Required Methods

fn save( &mut self, node: TreeNode, hash_pair: &(Hash, Hash) ) -> impl Future<Output = Result<(), Error>>

Save a hash pair for a node

fn sync(&mut self) -> impl Future<Output = Result<(), Error>>

sync to disk

Object Safety

This trait is not object safe.