Module iroh::blobs::store::bao_tree::io::outboard

Implementations of the outboard traits

A number of implementations for the sync and async outboard traits are provided. Implementations for in-memory outboards, for outboards where the data resides on disk, and a special implementation EmptyOutboard that just ignores all writes.

Structs

EmptyOutboard

An empty outboard, that just returns 0 hashes for all nodes.

PostOrderMemOutboard

A post order outboard that is optimized for memory storage.

PostOrderOutboard

A generic outboard in post order

PreOrderMemOutboard

A pre order outboard that is optimized for memory storage.

PreOrderOutboard

A generic outboard in pre order