Module iroh::blobs::store::bao_tree::io::sync

Sync IO operations

The traits to perform positioned io are re-exported from positioned-io.

Structs

DecodeResponseIter

Iterator that can be used to decode a response to a range request

Traits

CreateOutboard

Convenience trait to initialize an outboard from a data source.

Outboard

A binary merkle tree for blake3 hashes of a blob.

OutboardMut

A mutable outboard.

ReadAt

Trait for reading bytes at an offset.

Size

Trait to get the size in bytes of an I/O object.

WriteAt

Trait for writing bytes at an offset.

Functions

copy

Copy an outboard to another outboard.

decode_ranges

Decode a response into a file while updating an outboard.

encode_ranges

Encode ranges relevant to a query from a reader and outboard to a writer

encode_ranges_validated

Encode ranges relevant to a query from a reader and outboard to a writer

outboard

Compute the outboard for the given data.

outboard_post_order

Compute the post order outboard for the given data, writing into a io::Write

valid_outboard_ranges

Given just an outboard, compute all valid ranges.

valid_ranges

Given a data file and an outboard, compute all valid ranges.