Trait iroh::blobs::store::bao_tree::io::sync::ReadAt

pub trait ReadAt {
    // Required method
    fn read_at(&self, pos: u64, buf: &mut [u8]) -> Result<usize, Error>;

    // Provided method
    fn read_exact_at(&self, pos: u64, buf: &mut [u8]) -> Result<(), Error> { ... }
}

Trait for reading bytes at an offset.

Implementations should be able to read bytes without changing any sort of read position. Self should not change at all. Buffering reads is unlikely to be useful, since each time read_at() is called, the position may be completely different.

Examples

Read the fifth 512-byte sector of a file:

use std::fs::File;
use positioned_io::ReadAt;

// note that file does not need to be mut
let file = File::open("tests/pi.txt")?;

// read up to 512 bytes
let mut buf = [0; 512];
let bytes_read = file.read_at(2048, &mut buf)?;

Required Methods

fn read_at(&self, pos: u64, buf: &mut [u8]) -> Result<usize, Error>

Reads bytes from an offset in this source into a buffer, returning how many bytes were read.

This function may yield fewer bytes than the size of buf, if it was interrupted or hit the “end of file”.

See Read::read() for details.

Provided Methods

fn read_exact_at(&self, pos: u64, buf: &mut [u8]) -> Result<(), Error>

Reads the exact number of bytes required to fill buf from an offset.

Errors if the “end of file” is encountered before filling the buffer.

See Read::read_exact() for details.