Module iroh::blobs::store::bao_tree::iter

Iterators over BaoTree nodes

Range iterators take a reference to the ranges, and therefore require a lifetime parameter. They can be used without lifetime parameters using self referencing structs.

Structs

PostOrderChunkIter

Iterator over all chunks in a BaoTree in post-order.

PostOrderNodeIter

Iterator over all nodes in a tree in post-order.

PreOrderNodeIter

Iterator over all nodes in a BaoTree in pre-order.

PreOrderPartialChunkIterRef

Iterator over all nodes in a BaoTree in pre-order that overlap with a given chunk range.

ResponseIter

The owned version of ResponseIterRef.

ResponseIterRef

An iterator that produces chunks in pre order.

Enums

BaoChunk

A chunk describeds what to read or write next