Trait iroh::blobs::store::range_collections::range_set::RangeSetEntry

pub trait RangeSetEntry: Ord {
    // Required methods
    fn min_value() -> Self;
    fn is_min_value(&self) -> bool;
}

trait for types that can be entries of range sets

they must have an order and a minimum value.

Required Methods

fn min_value() -> Self

the minimum value for this type

fn is_min_value(&self) -> bool

checks if this is the minimum value

this is to be able to check for minimum without having to create a value

Object Safety

This trait is not object safe.