Struct iroh::blobs::store::range_collections::RangeSet

pub struct RangeSet<A>(/* private fields */)
where
    A: Array;

A set of non-overlapping ranges

let mut a: RangeSet2<i32> = RangeSet::from(10..);
let b: RangeSet2<i32> = RangeSet::from(1..5);

a |= b;
let r = !a;

A data structure to represent a set of non-overlapping ranges of element type T: RangeSetEntry. It uses a SmallVec<T> of sorted boundaries internally.

It can represent not just finite ranges but also ranges with unbounded end. Because it can represent infinite ranges, it can also represent the set of all elements, and therefore all boolean operations including negation.

Adjacent ranges will be merged.

It provides very fast operations for set operations (&, |, ^) as well as for intersection tests (is_disjoint, is_subset).

In addition to the fast set operations that produce a new range set, it also supports the equivalent in-place operations.

Complexity

Complexity is given separately for the number of comparisons and the number of copies, since sometimes you have a comparison operation that is basically free (any of the primitive types), whereas sometimes you have a comparison operation that is many orders of magnitude more expensive than a copy (long strings, arbitrary precision integers, …)

Number of comparisons

operation	best	worst	remark

negation	1 	O(N)

union	O(log(N))	O(N)	binary merge

intersection	O(log(N))	O(N)	binary merge

difference	O(log(N))	O(N)	binary merge

xor	O(log(N))	O(N)	binary merge

membership	O(log(N))	O(log(N))	binary search

is_disjoint	O(log(N))	O(N)	binary merge with cutoff

is_subset	O(log(N))	O(N)	binary merge with cutoff

Number of copies

For creating new sets, obviously there needs to be at least one copy for each element of the result set, so the complexity is always O(N). For in-place operations it gets more interesting. In case the number of elements of the result being identical to the number of existing elements, there will be no copies and no allocations.

E.g. if the result just has some of the ranges of the left hand side extended or truncated, but the same number of boundaries, there will be no allocations and no copies except for the changed boundaries themselves.

If the result has fewer boundaries than then lhs, there will be some copying but no allocations. Only if the result is larger than the capacity of the underlying vector of the lhs will there be allocations.

operation	best	worst

negation	1 	1

union	1	O(N)

intersection	1	O(N)

difference	1	O(N)

xor	1	O(N)

Testing

Testing is done by some simple smoke tests as well as quickcheck tests of the algebraic properties of the boolean operations.

Implementations

impl<T, A> RangeSet<A>
where A: Array<Item = T>,

pub fn new(boundaries: SmallVec<A>) -> Option<RangeSet<A>>
where <A as Array>::Item: Ord,

create a new range set from the given boundaries

This performs a check that the boundaries are strictly sorted. If you want to avoid this check, use new_unchecked (behind a feature flag because it is unsafe)

pub fn new_unchecked(boundaries: SmallVec<A>) -> RangeSet<A>

Create a new range set reference without checking that the boundaries are strictly sorted.

pub fn iter(&self) -> Iter<'_, T>

iterate over all ranges in this range set

pub fn into_inner(self) -> SmallVec<A>

get the boundaries in this range set as a SmallVec

impl<T, A> RangeSet<A>
where T: RangeSetEntry, A: Array<Item = T>,

pub fn empty() -> RangeSet<A>

the empty range set

pub fn all() -> RangeSet<A>

a range set containing all values

impl<T, A> RangeSet<A>
where T: RangeSetEntry + Clone, A: Array<Item = T>,

pub fn intersection_with(&mut self, that: &RangeSetRef<T>)

intersection in place

pub fn union_with(&mut self, that: &RangeSetRef<T>)

union in place

pub fn difference_with(&mut self, that: &RangeSetRef<T>)

difference in place

pub fn symmetric_difference_with(&mut self, that: &RangeSetRef<T>)

symmetric difference in place (xor)