Module iroh::client::blobs

API for blobs management.

The main entry point is the Client.

You obtain a Client via Iroh::blobs().

Interacting with the local blob store

Importing data

There are several ways to import data into the local blob store:

add_bytes imports in memory data.

add_stream imports data from a stream of bytes.

add_reader imports data from an async reader.

add_from_path imports data from a file.

The last method imports data from a file on the local filesystem. This is the most efficient way to import large amounts of data.

Exporting data

There are several ways to export data from the local blob store:

read_to_bytes reads data into memory.

read creates a reader to read data from.

export eports data to a file on the local filesystem.

Interacting with remote nodes

download downloads data from a remote node.

share allows creating a ticket to share data with a remote node.

Interacting with the blob store itself

These are more advanced operations that are usually not needed in normal operation.

consistency_check checks the internal consistency of the local blob store.

validate validates the locally stored data against their BLAKE3 hashes.

delete_blob deletes a blob from the local store.

Structs

AddOutcome

Outcome of a blob add operation.

AddProgress

Progress stream for blob add operations.

BlobInfo

Information about a complete blob.

Client

Iroh blobs client.

CollectionInfo

Information about a stored collection.

DownloadOptions

Options to configure a download request.

DownloadOutcome

Outcome of a blob download operation.

DownloadProgress

Progress stream for blob download operations.

ExportOutcome

Outcome of a blob export operation.

ExportProgress

Progress stream for blob export operations.

IncompleteBlobInfo

Information about an incomplete blob.

Reader

Data reader for a single blob.

Enums

BlobStatus

Status information about a blob.

DownloadMode

Set the mode for whether to directly start the download or add it to the download queue.

WrapOption

Whether to wrap the added data in a collection.