Module iroh::client::gossip

Gossip client.

The gossip client allows you to subscribe to gossip topics and send updates to them.

The main entry point is the Client.

You obtain a Client via Iroh::gossip().

The gossip API is extremely simple. You use subscribe to subscribe to a topic. This returns a sink to send updates to the topic and a stream of responses.

Client::subscribe_with_opts allows you to specify advanced options such as the buffer size.

Structs

Client

Iroh gossip client.

SubscribeOpts

Options for subscribing to a gossip topic.

SubscribeRequest

A request to the node to subscribe to gossip events.

Enums

SubscribeResponse

Events emitted from a gossip topic with a lagging notification.

SubscribeUpdate

Send a gossip message