iroh::discovery

Enum DiscoveryError

#[non_exhaustive]
pub enum DiscoveryError {
    NoServiceConfigured {
        backtrace: Option<Backtrace>,
        span_trace: SpanTrace,
    },
    NoResults {
        node_id: NodeId,
        backtrace: Option<Backtrace>,
        span_trace: SpanTrace,
    },
    User {
        provenance: &'static str,
        source: Box<dyn Error + Send + Sync + 'static>,
        backtrace: Option<Backtrace>,
        span_trace: SpanTrace,
    },
}

Discovery errors

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

NoServiceConfigured

Fields

backtrace: Option<Backtrace>

span_trace: SpanTrace

NoResults

Fields

node_id: NodeId

backtrace: Option<Backtrace>

span_trace: SpanTrace

User

Fields

provenance: &'static str

source: Box<dyn Error + Send + Sync + 'static>

backtrace: Option<Backtrace>

span_trace: SpanTrace

Implementations

impl DiscoveryError

pub fn from_err<T: Error + Send + Sync + 'static>( provenance: &'static str, source: T, ) -> Self

Creates a new user error from an arbitrary error type.

pub fn from_err_box( provenance: &'static str, source: Box<dyn Error + Send + Sync + 'static>, ) -> Self

Creates a new user error from an arbitrary boxed error type.