iroh::discovery::pkarr

Module dht

Available on crate feature discovery-pkarr-dht only.

Pkarr based endpoint discovery for iroh, supporting both relay servers and the DHT.

This module contains pkarr-based endpoint discovery for iroh which can use both pkarr relay servers as well as the Mainline DHT directly. See the pkarr module for an overview of pkarr.

Structs

Builder

Builder for DhtDiscovery.

DhtDiscovery

Pkarr Mainline DHT and relay server endpoint discovery.