iroh::discovery

Module static_provider

A static endpoint discovery to manually add endpoint addressing information.

Often an application might get endpoint addressing information out-of-band in an application-specific way. EndpointTicket’s are one common way used to achieve this. This “static” addressing information is often only usable for a limited time so needs to be able to be removed again once know it is no longer useful.

This is where the StaticProvider is useful: it allows applications to add and retract endpoint addressing information that is otherwise out-of-band to iroh.

Structs

StaticProvider

A static endpoint discovery to manually add endpoint addressing information.