iroh::discovery::static_provider
Struct StaticProvider

pub struct StaticProvider { /* private fields */ }

A static node discovery to manually add node addressing information.

Often an application might get node addressing information out-of-band in an application-specific way. NodeTicketโs are one common way used to achieve this. This โstaticโ addressing information is often only usable for a limited time so needs to be able to be removed again once know it is no longer useful.

This is where the StaticProvider is useful: it allows applications to add and retract node addressing information that is otherwise out-of-band to iroh.

Examples

use iroh::{discovery::static_provider::StaticProvider, Endpoint, NodeAddr};
use iroh_base::SecretKey;

// Create the discovery service and endpoint.
let discovery = StaticProvider::new();

let _ep = Endpoint::builder()
    .add_discovery({
        let discovery = discovery.clone();
        move |_| Some(discovery)
    })
    .bind()
    .await?;

// Sometime later add a RelayUrl for a fake NodeId.
let node_id = SecretKey::from_bytes(&[0u8; 32]).public(); // Do not use fake secret keys!
// You can pass either `NodeInfo` or `NodeAddr` to `add_node_info`.
discovery.add_node_info(
    NodeAddr {
        node_id,
        relay_url: Some("https://example.com".parse()?),
        direct_addresses: Default::default(),
    },
);

Implementations

impl StaticProvider

pub const PROVENANCE: &'static str = "static_discovery"

The provenance string for this discovery implementation.

This is mostly used for debugging information and allows understanding the origin of addressing information used by an iroh Endpoint.

pub fn new() -> Self

Creates a new static discovery instance.

pub fn from_node_info( infos: impl IntoIterator<Item = impl Into<NodeInfo>>, ) -> Self

Creates a static discovery instance from node addresses.

Examples

use std::{net::SocketAddr, str::FromStr};

use iroh::{discovery::static_provider::StaticProvider, Endpoint, NodeAddr};

// get addrs from somewhere
let addrs = get_addrs();

// create a StaticProvider from the list of addrs.
let discovery = StaticProvider::from_node_info(addrs);
// create an endpoint with the discovery
let endpoint = Endpoint::builder()
    .add_discovery(|_| Some(discovery))
    .bind()
    .await?;

pub fn set_node_info(&self, node_info: impl Into<NodeInfo>) -> Option<NodeData>

Sets node addressing information for the given node ID.

This will completely overwrite any existing info for the node.

Returns the NodeData of the previous entry, or None if there was no previous entry for this node ID.

pub fn add_node_info(&self, node_info: impl Into<NodeInfo>)

Augments node addressing information for the given node ID.

The provided addressing information is combined with the existing info in the static provider. Any new direct addresses are added to those already present while the relay URL is overwritten.

pub fn get_node_info(&self, node_id: NodeId) -> Option<NodeInfo>

Returns node addressing information for the given node ID.

pub fn remove_node_info(&self, node_id: NodeId) -> Option<NodeInfo>

Removes all node addressing information for the given node ID.

Any removed information is returned.