iroh::discovery

Struct EndpointInfo

pub struct EndpointInfo {
    pub endpoint_id: PublicKey,
    pub data: EndpointData,
}

Information about an endpoint that may be published to and resolved from discovery services.

This struct couples a EndpointId with its associated EndpointData.

Fields
endpoint_id: PublicKey

The EndpointId of the endpoint this is about.

data: EndpointData

The information published about the endpoint.

Implementations

impl EndpointInfo

pub fn new(endpoint_id: PublicKey) -> EndpointInfo

Creates a new EndpointInfo with an empty EndpointData.

pub fn from_parts(endpoint_id: PublicKey, data: EndpointData) -> EndpointInfo

Creates a new EndpointInfo from its parts.

pub fn with_relay_url(self, relay_url: Option<RelayUrl>) -> EndpointInfo

Sets the relay URL and returns the updated endpoint info.

pub fn with_ip_addrs(self, addrs: BTreeSet<SocketAddr>) -> EndpointInfo

Sets the IP based addresses and returns the updated endpoint info.

pub fn with_user_data(self, user_data: Option<UserData>) -> EndpointInfo

Sets the user-defined data and returns the updated endpoint info.

pub fn to_endpoint_addr(&self) -> EndpointAddr

Converts into a EndpointAddr by cloning the needed fields.

pub fn into_endpoint_addr(self) -> EndpointAddr

Converts into a EndpointAddr without cloning.

pub fn from_txt_lookup( domain_name: String, lookup: impl Iterator<Item = TxtRecordData>, ) -> Result<EndpointInfo, ParseError>

Parses a EndpointInfo from DNS TXT lookup.

pub fn from_pkarr_signed_packet( packet: &SignedPacket, ) -> Result<EndpointInfo, ParseError>

Parses a EndpointInfo from a [pkarr::SignedPacket].

pub fn to_pkarr_signed_packet( &self, secret_key: &SecretKey, ttl: u32, ) -> Result<SignedPacket, EncodingError>

Creates a [pkarr::SignedPacket].

This constructs a DNS packet and signs it with a SecretKey.

pub fn to_txt_strings(&self) -> Vec<String>

Converts into a list of {key}={value} strings.