iroh::discovery
Trait IntoDiscovery

pub trait IntoDiscovery:
    Send
    + Sync
    + 'static {
    // Required method
    fn into_discovery(self, endpoint: &Endpoint) -> Result<impl Discovery>;
}

Trait for structs that can be converted into Discovery.

This trait is implemented on builders for discovery services. Any type that implements this trait can be added as a discovery service in Builder::add_discovery.

Any type that implements Discovery also implements IntoDiscovery.

Required Methods

fn into_discovery(self, endpoint: &Endpoint) -> Result<impl Discovery>

Turns this struct into a boxed Discovery.

The discovery service will be used by passed Endpoint. If a discovery service needs an endpoint or information from it like its secret key, it can access or clone the endpoint from here.

If an error is returned, building the endpoint will fail with this error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.