Module iroh::dns

This module exports a DNS resolver, which is also the default resolver used in the crate::Endpoint if no custom resolver is configured.

It also exports ResolverExt: A extension trait over DnsResolver to perform DNS queries by ipv4, ipv6, name and node_id. See the node_info module documentation for details on how iroh node records are structured.

Modules

node_info

Support for handling DNS resource records for dialing by NodeId.

Traits

ResolverExt

Extension trait to DnsResolver.

Functions

default_resolver

Get a reference to the default DNS resolver.

resolver

Get the DNS resolver used within iroh.

Type Aliases

DnsResolver

The DNS resolver type used throughout iroh.