iroh::dns::node_info
Struct TxtAttrs

pub struct TxtAttrs<T> { /* private fields */ }

Attributes parsed from IROH_TXT_NAME TXT records.

This struct is generic over the key type. When using with String, this will parse all attributes. Can also be used with an enum, if it implements FromStr and Display.

Implementations

impl<T: FromStr + Display + Hash + Ord> TxtAttrs<T>

pub fn from_parts( node_id: NodeId, pairs: impl Iterator<Item = (T, String)>, ) -> Self

Creates TxtAttrs from a node id and an iterator of key-value pairs.

pub fn from_strings( node_id: NodeId, strings: impl Iterator<Item = String>, ) -> Result<Self>

Creates TxtAttrs from a node id and an iterator of “{key}={value}” strings.

pub async fn lookup_by_id( resolver: &TokioResolver, node_id: &NodeId, origin: &str, ) -> Result<Self>

Looks up attributes by NodeId and origin domain.

pub async fn lookup_by_name( resolver: &TokioResolver, name: &str, ) -> Result<Self>

Looks up attributes by DNS name.

pub fn attrs(&self) -> &BTreeMap<T, Vec<String>>

Returns the parsed attributes.

pub fn node_id(&self) -> NodeId

Returns the node id.

pub fn from_pkarr_signed_packet(packet: &SignedPacket) -> Result<Self>

Parses a [pkarr::SignedPacket].

pub fn from_hickory_lookup(lookup: TxtLookup) -> Result<Self>

Parses a TXT records lookup.

pub fn to_hickory_records( &self, origin: &str, ttl: u32, ) -> Result<impl Iterator<Item = Record> + '_>

Converts to a list of [hickory_resolver::proto::rr::Record] resource records.

pub fn to_pkarr_signed_packet( &self, secret_key: &SecretKey, ttl: u32, ) -> Result<SignedPacket>

Creates a [pkarr::SignedPacket]

This constructs a DNS packet and signs it with a SecretKey.