iroh::dns

Trait DnsResolverExt

pub trait DnsResolverExt {
    // Required methods
    fn lookup_endpoint_by_id(
        &self,
        endpoint_id: &PublicKey,
        origin: &str,
    ) -> impl Future<Output = Result<EndpointInfo, LookupError>> + Send;
    fn lookup_endpoint_by_domain_name(
        &self,
        name: &str,
    ) -> impl Future<Output = Result<EndpointInfo, LookupError>> + Send;
    fn lookup_endpoint_by_domain_name_staggered(
        &self,
        name: &str,
        delays_ms: &[u64],
    ) -> impl Future<Output = Result<EndpointInfo, StaggeredError<LookupError>>> + Send;
    fn lookup_endpoint_by_id_staggered(
        &self,
        endpoint_id: &PublicKey,
        origin: &str,
        delays_ms: &[u64],
    ) -> impl Future<Output = Result<EndpointInfo, StaggeredError<LookupError>>> + Send;
}

Available on non-wasm_browser only.

Extension trait adding iroh endpoint lookup methods to DnsResolver.

Required Methods

fn lookup_endpoint_by_id( &self, endpoint_id: &PublicKey, origin: &str, ) -> impl Future<Output = Result<EndpointInfo, LookupError>> + Send

Looks up endpoint info by EndpointId and origin domain name.

To lookup endpoints that published their endpoint info to the DNS servers run by n0, pass N0_DNS_ENDPOINT_ORIGIN_PROD as origin.

fn lookup_endpoint_by_domain_name( &self, name: &str, ) -> impl Future<Output = Result<EndpointInfo, LookupError>> + Send

Looks up endpoint info by DNS name.

fn lookup_endpoint_by_domain_name_staggered( &self, name: &str, delays_ms: &[u64], ) -> impl Future<Output = Result<EndpointInfo, StaggeredError<LookupError>>> + Send

Looks up endpoint info by DNS name in a staggered fashion.

fn lookup_endpoint_by_id_staggered( &self, endpoint_id: &PublicKey, origin: &str, delays_ms: &[u64], ) -> impl Future<Output = Result<EndpointInfo, StaggeredError<LookupError>>> + Send

Looks up endpoint info by EndpointId and origin domain name in a staggered fashion.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.