Crate iroh::docs

Multi-dimensional key-value documents with an efficient synchronization protocol

The crate operates on Replicas. A replica contains an unlimited number of Entries. Each entry is identified by a key, its author, and the replica’s namespace. Its value is the 32-byte BLAKE3 hash of the entry’s content data, the size of this content data, and a timestamp. The content data itself is not stored or transferred through a replica.

All entries in a replica are signed with two keypairs:

The NamespaceSecret key, as a token of write capability. The public key is the NamespaceId, which also serves as the unique identifier for a replica.

The Author key, as a proof of authorship. Any number of authors may be created, and their semantic meaning is application-specific. The public key of an author is the AuthorId.

Replicas can be synchronized between peers by exchanging messages. The synchronization algorithm is based on a technique called range-based set reconciliation, based on this paper by Aljoscha Meyer:

Range-based set reconciliation is a simple approach to efficiently compute the union of two sets over a network, based on recursively partitioning the sets and comparing fingerprints of the partitions to probabilistically detect whether a partition requires further work.

The crate exposes a generic storage interface. There is an implementation of this interface, store::fs::Store, that can be used either in-memory or in persistent, file-based mode.

Both modes make use of [redb], an embedded key-value store. When used in-memory, the store is backed by a Vec<u8>. When used in persistent mode, the store is backed by a single file on disk.

Modules

actor

This contains an actor spawned on a separate thread to process replica and store operations.

engine

Handlers and actors to for live syncing replicas.

metrics

Metrics for iroh-docs

net

Network implementation of the iroh-docs protocol

store

Storage trait and implementation for iroh-docs documents

sync

API for iroh-docs replicas

Structs

Author

Author key to insert entries in a crate::Replica

AuthorHeads

Timestamps of the latest entry for each author.

AuthorId

AuthorPublicKey in bytes

AuthorPublicKey

Identifier for an Author

DocTicket

Contains both a key (either secret or public) to a document, and a list of peers to join.

Entry

A single entry in a Replica

EntrySignature

Signature over an entry.

NamespaceId

NamespacePublicKey in bytes

NamespacePublicKey

The corresponding [VerifyingKey] for a NamespaceSecret. It is used as an identifier, and can be used to verify Signatures.

NamespaceSecret

Namespace key of a crate::Replica.

ReadOnly

Error that occurs trying to access the NamespaceSecret of a read-only Capability.

Record

The data part of an entry in a Replica.

RecordIdentifier

The identifier of a record.

Replica

Local representation of a mutable, synchronizable key-value store.

ReplicaInfo

In memory information about an open replica.

SignedEntry

A signed entry.

SyncOutcome

Outcome of a sync operation.

Enums

Capability

The capability of the namespace.

CapabilityError

Errors for capability operations

CapabilityKind

Kind of capability of the namespace.

ContentStatus

Whether the content status is available on a node.

Event

Event emitted by sync when entries are added.

InsertError

Error emitted when inserting entries into a Replica failed

InsertOrigin

Whether an entry was inserted locally or by a remote peer.

ValidationFailure

Reason why entry validation failed

Constants

MAX_TIMESTAMP_FUTURE_SHIFT

Max time in the future from our wall clock time that we accept entries for. Value is 10 minutes.

Type Aliases

ContentStatusCallback

Callback that may be set on a replica to determine the availability status for a content hash.

PeerIdBytes

Byte representation of a PeerId from iroh-net.

ProtocolMessage

Protocol message for the set reconciliation protocol.