Enum iroh::docs::cli::DocCommands

pub enum DocCommands {

    Switch {
        id: NamespaceId,
    },
    Create {
        switch: bool,
    },
    Join {
        ticket: DocTicket,
        switch: bool,
    },
    List,
    Share {
        doc: Option<NamespaceId>,
        mode: ShareMode,
        addr_options: AddrInfoOptions,
    },
    Set {
        doc: Option<NamespaceId>,
        author: Option<AuthorId>,
        key: String,
        value: String,
    },
    DlPolicy(DlPolicyCmd),
    Get {
        doc: Option<NamespaceId>,
        key: String,
        prefix: bool,
        author: Option<AuthorId>,
        mode: DisplayContentMode,
    },
    Del {
        doc: Option<NamespaceId>,
        author: Option<AuthorId>,
        prefix: String,
    },
    Keys {
        doc: Option<NamespaceId>,
        author: Option<AuthorId>,
        prefix: Option<String>,
        sort: Sorting,
        desc: bool,
        mode: DisplayContentMode,
    },
    Import {
        doc: Option<NamespaceId>,
        author: Option<AuthorId>,
        prefix: Option<String>,
        path: String,
        in_place: bool,
        no_prompt: bool,
    },
    Export {
        doc: Option<NamespaceId>,
        key: String,
        out: String,
    },
    Watch {
        doc: Option<NamespaceId>,
    },
    Leave {
        doc: Option<NamespaceId>,
    },
    Drop {
        doc: Option<NamespaceId>,
    },

}

Possible Document commands.

Variants

Switch

Set the active document (only works within the Iroh console).

Fields

id: NamespaceId

Create

Create a new document.

Fields

switch: bool

Switch to the created document (only in the Iroh console).

Join

Join a document from a ticket.

Fields

ticket: DocTicket

switch: bool

Switch to the joined document (only in the Iroh console).

List

List documents.

Share

Share a document with peers.

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

mode: ShareMode

The sharing mode.

addr_options: AddrInfoOptions

Options to configure the address information in the generated ticket.

Use relay-and-addresses in networks with no internet connectivity.

Set

Set an entry in a document.

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

author: Option<AuthorId>

Author of the entry.

Required unless the author is set through the IROH_AUTHOR environment variable. Within the Iroh console, the active author can also set with author switch.

key: String

Key to the entry (parsed as UTF-8 string).

value: String

Content to store for this entry (parsed as UTF-8 string)

DlPolicy(DlPolicyCmd)

Set the download policies for a document.

Get

Get entries in a document.

Shows the author, content hash and content length for all entries for this key.

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

key: String

Key to the entry (parsed as UTF-8 string).

prefix: bool

If true, get all entries that start with KEY.

author: Option<AuthorId>

Filter by author.

mode: DisplayContentMode

How to show the contents of the key.

Del

Delete all entries below a key prefix.

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

author: Option<AuthorId>

Author of the entry.

Required unless the author is set through the IROH_AUTHOR environment variable. Within the Iroh console, the active author can also set with author switch.

prefix: String

Prefix to delete. All entries whose key starts with or is equal to the prefix will be deleted.

Keys

List all keys in a document.

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

author: Option<AuthorId>

Filter by author.

prefix: Option<String>

Optional key prefix (parsed as UTF-8 string)

sort: Sorting

How to sort the entries

desc: bool

Sort in descending order

mode: DisplayContentMode

How to show the contents of the keys.

Import

Import data into a document

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also be set with doc switch.

author: Option<AuthorId>

Author of the entry.

Required unless the author is set through the IROH_AUTHOR environment variable. Within the Iroh console, the active author can also be set with author switch.

prefix: Option<String>

Prefix to add to imported entries (parsed as UTF-8 string). Defaults to no prefix

path: String

Path to a local file or directory to import

Pathnames will be used as the document key

in_place: bool

If true, don’t copy the file into iroh, reference the existing file instead

Moving a file imported with in-place will result in data corruption

no_prompt: bool

When true, you will not get a prompt to confirm you want to import the files

Export

Export the most recent data for a key from a document

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also be set with doc switch.

key: String

Key to the entry (parsed as UTF-8 string)

When just the key is present, will export the latest entry for that key.

out: String

Path to export to

Watch

Watch for changes and events on a document

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

Leave

Stop syncing a document.

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

Drop

Delete a document from the local node.

This is a destructive operation. Both the document secret key and all entries in the document will be permanently deleted from the node’s storage. Content blobs will be deleted through garbage collection unless they are referenced from another document or tag.

Fields

doc: Option<NamespaceId>

Document to operate on.

Required unless the document is set through the IROH_DOC environment variable. Within the Iroh console, the active document can also set with doc switch.

Implementations

impl DocCommands

pub async fn run( self, docs: &Client, blobs: &Client, env: &ConsoleEnv ) -> Result<(), Error>

Runs the document command given the iroh client and the console environment.