Module iroh::docs::engine

Available on crate feature engine only.

Handlers and actors to for live syncing replicas.

crate::Replica is also called documents here.

Structs

DefaultAuthor

Persistent default author for a docs engine.

Engine

The sync engine coordinates actors that manage open documents, set-reconciliation syncs with peers and a gossip swarm for each syncing document.

SyncEvent

Event emitted when a sync operation completes

Enums

DefaultAuthorStorage

Where to persist the default author.

LiveEvent

Events informing about actions of the live sync progress.

Origin

Why we performed a sync exchange

SyncReason

Why we started a sync request

Functions

entry_to_content_status

Converts an EntryStatus into a [‘ContentStatus’].