iroh::endpoint

Enum ConnectionError

pub enum ConnectionError {
    VersionMismatch,
    TransportError(Error),
    ConnectionClosed(ConnectionClose),
    ApplicationClosed(ApplicationClose),
    Reset,
    TimedOut,
    LocallyClosed,
    CidsExhausted,
}

Reasons why a connection might be lost

Variants

VersionMismatch

The peer doesnโt implement any supported version

TransportError(Error)

The peer violated the QUIC specification as understood by this implementation

ConnectionClosed(ConnectionClose)

The peerโs QUIC stack aborted the connection automatically

ApplicationClosed(ApplicationClose)

The peer closed the connection

Reset

The peer is unable to continue processing this connection, usually due to having restarted

TimedOut

Communication with the peer has lapsed for longer than the negotiated idle timeout

If neither side is sending keep-alives, a connection will time out after a long enough idle period even if the peer is still reachable. See also TransportConfig::max_idle_timeout() and TransportConfig::keep_alive_interval().

LocallyClosed

The local application closed the connection

CidsExhausted

The connection could not be created because not enough of the CID space is available

Try using longer connection IDs.