iroh::endpoint

Module quic

Exporting and encapsulating structs from quinn

Co-locates all iroh-quinn exports.

There are some structs that we use in particular ways, where we would like to limit or expand how those structs are used in iroh. By encapsulating them we can ensure the functionality needed to make iroh work.

Structs

AcceptBi

Future produced by [Connection::accept_bi]

AcceptUni

Future produced by [Connection::accept_uni]

AckFrequencyConfig

Parameters for controlling the peer’s acknowledgement frequency

ApplicationClose

Reason given by an application for closing the connection

Chunk

A chunk of data from the receive stream

ClosedStream

Error indicating that a stream has not been opened or has already been finished or reset

ConnectionClose

Reason given by the transport for closing the connection

ConnectionStats

Connection statistics

CryptoError

Generic crypto errors

ExportKeyingMaterialError

Error returned by [Session::export_keying_material].

FrameStats

Number of frames transmitted or received of each frame type

IdleTimeout

Maximum duration of inactivity to accept before timing out the connection

MtuDiscoveryConfig

Parameters governing MTU discovery.

OpenBi

Future produced by [Connection::open_bi]

OpenUni

Future produced by [Connection::open_uni]

PathStats

Statistics related to a transmission path

QuicTransportConfig

Parameters governing the core QUIC state machine

ReadDatagram

Future produced by [Connection::read_datagram]

RecvStream

A stream that can only be used to receive data

RetryError

Error for attempting to retry an [Incoming] which already bears a token from a previous retry

SendStream

A stream that can only be used to send data

ServerConfig

Parameters governing incoming connections

StreamId

Identifier for a stream within a particular connection

TransportError

Transport-level errors occur when a peer violates the protocol specification

TransportErrorCode

Transport-level error code

UdpStats

Statistics about UDP datagrams transmitted or received on a connection

UnsupportedVersion

Error indicating that the specified QUIC version is not supported

VarInt

An integer less than 2^62

VarIntBoundsExceeded

Error returned when constructing a VarInt from a value >= 2^62

WeakConnectionHandle

A handle to some connection internals, use with care.

Written

Indicates how many bytes and chunks had been transferred in a write operation

Enums

ConnectionError

Reasons why a connection might be lost

QuicTransportConfigError

ReadError

Errors that arise from reading from a stream.

ReadExactError

Errors that arise from reading from a stream.

ReadToEndError

Errors from RecvStream::read_to_end

ResetError

Errors that arise while waiting for a stream to be reset

SendDatagramError

Errors that can arise when sending a datagram

StoppedError

Errors that arise while monitoring for a send stream stop from the peer

WriteError

Errors that arise from writing to a stream

Traits

AeadKey

A key for sealing data with AEAD-based algorithms

Controller

Common interface for different congestion controllers

ControllerFactory

Constructs controllers on demand

CryptoServerConfig

Server-side configuration for the crypto protocol

HandshakeTokenKey

A pseudo random key for HKDF