iroh::endpoint

Struct IncomingZeroRttConnection

pub struct IncomingZeroRttConnection { /* private fields */ }

A QUIC connection on the server-side that can possibly accept 0-RTT data.

It is very similar to a Connection, but the IncomingZeroRttConnection::remote_id and IncomingZeroRttConnection::alpn may not be set yet, since the handshake has not necessarily occurred yet.

If the IncomingZeroRttConnection has rejected 0-RTT or does not have enough information to accept 0-RTT, any received 0-RTT packets will simply be dropped before reaching any receive streams.

Any streams that are created to send or receive data can continue to be used even after the handshake has completed and we are no longer in a 0-RTT situation.

Use the IncomingZeroRttConnection::handshake_completed method to get a Connection from a IncomingZeroRttConnection. This waits until 0-RTT connection has completed the handshake and can now confidently derive the ALPN and the EndpointId of the remote endpoint.

Implementations

impl IncomingZeroRttConnection

pub async fn handshake_completed( self, ) -> Result<Connection, AuthenticationError>

Waits until the full handshake occurs and then returns a Connection.

This may fail with AuthenticationError::ConnectionError, if there was some general failure with the connection, such as a network timeout since we accepted the connection.

This may fail with other AuthenticationErrors, if the other side doesn’t use the right TLS authentication, which usually every iroh endpoint uses and requires.

Thus, those errors should only occur if someone connects to you with a modified iroh endpoint or with a plain QUIC client.

pub fn open_uni(&self) -> OpenUni<'_>

Initiates a new outgoing unidirectional stream.

Streams are cheap and instantaneous to open unless blocked by flow control. As a consequence, the peer won’t be notified that a stream has been opened until the stream is actually used.

pub fn open_bi(&self) -> OpenBi<'_>

Initiates a new outgoing bidirectional stream.

Streams are cheap and instantaneous to open unless blocked by flow control. As a consequence, the peer won’t be notified that a stream has been opened until the stream is actually used. Calling open_bi then waiting on the RecvStream without writing anything to SendStream will never succeed.

pub fn accept_uni(&self) -> AcceptUni<'_>

Accepts the next incoming uni-directional stream.

pub fn accept_bi(&self) -> AcceptBi<'_>

Accept the next incoming bidirectional stream.

Important Note: The peer that calls open_bi must write to its SendStream before the peer Connection is able to accept the stream using accept_bi(). Calling open_bi then waiting on the RecvStream without writing anything to the connected SendStream will never succeed.

pub fn read_datagram(&self) -> ReadDatagram<'_>

Receives an application datagram.

pub async fn closed(&self) -> ConnectionError

Wait for the connection to be closed for any reason.

Despite the return type’s name, closed connections are often not an error condition at the application layer. Cases that might be routine include ConnectionError::LocallyClosed and ConnectionError::ApplicationClosed.

pub fn close_reason(&self) -> Option<ConnectionError>

If the connection is closed, the reason why.

Returns None if the connection is still open.

pub fn close(&self, error_code: VarInt, reason: &[u8])

Closes the connection immediately.

Pending operations will fail immediately with ConnectionError::LocallyClosed. No more data is sent to the peer and the peer may drop buffered data upon receiving the CONNECTION_CLOSE frame.

error_code and reason are not interpreted, and are provided directly to the peer.

reason will be truncated to fit in a single packet with overhead; to improve odds that it is preserved in full, it should be kept under 1KiB.

Gracefully closing a connection

Only the peer last receiving application data can be certain that all data is delivered. The only reliable action it can then take is to close the connection, potentially with a custom error code. The delivery of the final CONNECTION_CLOSE frame is very likely if both endpoints stay online long enough, calling Endpoint::close will wait to provide sufficient time. Otherwise, the remote peer will time out the connection, provided that the idle timeout is not disabled.

The sending side can not guarantee all stream data is delivered to the remote application. It only knows the data is delivered to the QUIC stack of the remote endpoint. Once the local side sends a CONNECTION_CLOSE frame in response to calling close the remote endpoint may drop any data it received but is as yet undelivered to the application, including data that was acknowledged as received to the local endpoint.

pub fn send_datagram(&self, data: Bytes) -> Result<(), SendDatagramError>

Transmits data as an unreliable, unordered application datagram.

Application datagrams are a low-level primitive. They may be lost or delivered out of order, and data must both fit inside a single QUIC packet and be smaller than the maximum dictated by the peer.

pub fn max_datagram_size(&self) -> Option<usize>

Computes the maximum size of datagrams that may be passed to send_datagram.

Returns None if datagrams are unsupported by the peer or disabled locally.

This may change over the lifetime of a connection according to variation in the path MTU estimate. The peer can also enforce an arbitrarily small fixed limit, but if the peer’s limit is large this is guaranteed to be a little over a kilobyte at minimum.

Not necessarily the maximum size of received datagrams.

pub fn datagram_send_buffer_space(&self) -> usize

Bytes available in the outgoing datagram buffer.

When greater than zero, calling send_datagram with a datagram of at most this size is guaranteed not to cause older datagrams to be dropped.

pub fn rtt(&self) -> Duration

Current best estimate of this connection’s latency (round-trip-time).

pub fn stats(&self) -> ConnectionStats

Returns connection statistics.

pub fn congestion_state(&self) -> Box<dyn Controller>

Current state of the congestion control algorithm, for debugging purposes.

pub fn handshake_data(&self) -> Option<Box<dyn Any>>

Parameters negotiated during the handshake.

Guaranteed to return Some on fully established connections or after Connecting::handshake_data() succeeds. See that method’s documentations for details on the returned value.

pub fn alpn(&self) -> Option<Vec<u8>>

Extracts the ALPN protocol from the peer’s handshake data.

pub fn peer_identity(&self) -> Option<Box<dyn Any>>

Cryptographic identity of the peer.

The dynamic type returned is determined by the configured Session. For the default rustls session, the return value can be downcast to a Vec<[rustls::pki_types::CertificateDer]>

pub fn remote_id(&self) -> Result<EndpointId, RemoteEndpointIdError>

Returns the EndpointId from the peer’s TLS certificate.

The PublicKey of an endpoint is also known as an EndpointId. This PublicKey is included in the TLS certificate presented during the handshake when connecting. This function allows you to get the EndpointId of the remote endpoint of this connection.

pub fn stable_id(&self) -> usize

A stable identifier for this connection.

Peer addresses and connection IDs can change, but this value will remain fixed for the lifetime of the connection.

pub fn export_keying_material( &self, output: &mut [u8], label: &[u8], context: &[u8], ) -> Result<(), ExportKeyingMaterialError>

Derives keying material from this connection’s TLS session secrets.

When both peers call this method with the same label and context arguments and output buffers of equal length, they will get the same sequence of bytes in output. These bytes are cryptographically strong and pseudorandom, and are suitable for use as keying material.

See RFC5705 for more information.

pub fn set_max_concurrent_uni_streams(&self, count: VarInt)

Modifies the number of unidirectional streams that may be concurrently opened.

No streams may be opened by the peer unless fewer than count are already open. Large counts increase both minimum and worst-case memory consumption.

pub fn set_receive_window(&self, receive_window: VarInt)

See quinn_proto::TransportConfig::receive_window.

pub fn set_max_concurrent_bi_streams(&self, count: VarInt)

Modifies the number of bidirectional streams that may be concurrently opened.

No streams may be opened by the peer unless fewer than count are already open. Large counts increase both minimum and worst-case memory consumption.