iroh::endpoint

Type Alias IncomingZeroRttConnection

pub type IncomingZeroRttConnection = Connection<IncomingZeroRtt>;

A QUIC connection on the server-side that can possibly accept 0-RTT data.

It is very similar to a Connection, but the IncomingZeroRttConnection::remote_id and IncomingZeroRttConnection::alpn may not be set yet, since the handshake has not necessarily occurred yet.

If the IncomingZeroRttConnection has rejected 0-RTT or does not have enough information to accept 0-RTT, any received 0-RTT packets will simply be dropped before reaching any receive streams.

Any streams that are created to send or receive data can continue to be used even after the handshake has completed and we are no longer in a 0-RTT situation.

Use the IncomingZeroRttConnection::handshake_completed method to get a Connection from a IncomingZeroRttConnection. This waits until 0-RTT connection has completed the handshake and can now confidently derive the ALPN and the EndpointId of the remote endpoint.

Aliased Type

pub struct IncomingZeroRttConnection { /* private fields */ }