iroh

Module endpoint_id_resolution

Endpoint ID to address resolution .

To connect to an iroh endpoint a EndpointAddr is needed, which may contain a RelayUrl or one or more direct addresses in addition to the EndpointId.

Since there is a conversion from EndpointId to EndpointAddr, you can also use connect directly with a EndpointId.

For this to work however, the endpoint has to get the addressing information by other means.

Endpoint ID resolution is an automated system for an Endpoint to retrieve this addressing information. Each iroh endpoint will automatically publish their own addressing information. Usually this means publishing which RelayUrl to use for their EndpointId, but they could also publish their direct addresses.

The EndpointIdResolution trait is used to define endpoint ID resolution. This allows multiple implementations to co-exist because there are many possible ways to implement this. Each Endpoint can use the endpoint ID resolution mechanisms most suitable to the application. The Builder::endpoint_id_resolution method is used to add an endpoint ID resolution mechanism to an Endpoint.

Some generally useful EIR implementations are provided:

StaticProvider which allows application to add and remove out-of-band addressing information.

The DnsEndpointIdResolution which performs lookups via the standard DNS systems. To publish to this DNS server a PkarrPublisher is needed. Number 0 runs a public instance of a PkarrPublisher with attached DNS server which is globally available and a reliable default choice.

The PkarrResolver which can perform lookups from designated pkarr relay servers using HTTP.

MdnsEndpointIdResolution: mdns::MdnsEndpointIdResolution which uses the crate swarm-discovery, an opinionated mDNS implementation, to discover endpoints on the local network.

The DhtEndpointIdResolution also uses the pkarr system but can also publish and lookup records to/from the Mainline DHT.

To use multiple EIR systems simultaneously you can call Builder::endpoint_id_resolution. This will use ConcurrentEndpointIdResolution under the hood, which performs lookups to all EIR systems at the same time.

Builder::endpoint_id_resolution takes any type that implements IntoEndpointIdResolution. You can implement that trait on a builder struct if your EIR service needs information from the endpoint it is mounted on. After endpoint construction, your EIR service is built by calling IntoEndpointIdResolution::into_endpoint_id_resolution, passing the finished Endpoint to your builder.

If your EIR service does not need any information from its endpoint, you can pass the EIR service directly to Builder::endpoint_id_resolution: All types that implement EndpointIdResolution also have a blanket implementation of IntoEndpointIdResolution.

Examples

A very common setup is to enable DNS EIR, which needs to be done in two parts as a PkarrPublisher and DnsEndpointIdResolution:

use iroh::{
    Endpoint, SecretKey,
    endpoint::RelayMode,
    endpoint_id_resolution::{dns::DnsEndpointIdResolution, pkarr::PkarrPublisher},
};

let ep = Endpoint::empty_builder(RelayMode::Default)
    .endpoint_id_resolution(PkarrPublisher::n0_dns())
    .endpoint_id_resolution(DnsEndpointIdResolution::n0_dns())
    .bind()
    .await?;

To also enable MdnsEndpointIdResolution it can be added as another service.

#[cfg(feature = "mdns")]
let ep = Endpoint::empty_builder(RelayMode::Default)
    .endpoint_id_resolution(PkarrPublisher::n0_dns())
    .endpoint_id_resolution(DnsEndpointIdResolution::n0_dns())
    .endpoint_id_resolution(MdnsEndpointIdResolution::builder())
    .bind()
    .await?;

Modules

dnsNon-wasm_browser

DNS endpoint discovery for iroh

mdnsmdns

An endpoint ID resolution service that uses an mdns-like service to discover and resolve the addresses of local endpoints.

pkarr

An endpoint ID resolution service which publishes and resolves endpoint information using a pkarr relay.

static_provider

A static endpoint id resolution system to manually add endpoint addressing information.

Structs

ConcurrentEndpointIdResolution

An endpoint ID resolution service that combines multiple EIR sources.

EndpointData

Data about an endpoint that may be published to and resolved from discovery services.

EndpointIdResolutionItem

Endpoint ID resolution results from EndpointIdResolution services.

EndpointInfo

Information about an endpoint that may be published to and resolved from discovery services.

UserData

Under the hood this is a UTF-8 String is no longer than UserData::MAX_LENGTH bytes.

Enums

EndpointIdResolutionError

EndpointIdResolution errors

IntoEndpointIdResolutionError

IntoEndpointIdResolution errors

ParseError

Traits

EndpointIdResolution

Endpoint ID resolution for super::Endpoint.

IntoEndpointIdResolution

Trait for structs that can be converted into EndpointIdResolution systems.