iroh::endpoint_id_resolution

Trait EndpointIdResolution

pub trait EndpointIdResolution:
    Debug
    + Send
    + Sync
    + 'static {
    // Provided methods
    fn publish(&self, _data: &EndpointData) { ... }
    fn resolve(
        &self,
        _endpoint_id: EndpointId,
    ) -> Option<BoxStream<Result<EndpointIdResolutionItem, EndpointIdResolutionError>>> { ... }
}

Endpoint ID resolution for super::Endpoint.

This trait defines publishing and resolving addressing information for a EndpointId. This enables connecting to other endpoints with only knowing the EndpointId, by using this EndpointIdResolution system to look up the actual addressing information. It is common for implementations to require each endpoint to publish their own information before it can be looked up by other endpoints.

The published addressing information can include both a RelayUrl and/or direct addresses. See EndpointData for details.

To allow for EIR, the super::Endpoint will call publish whenever EIR information changes. If an EIR mechanism requires a periodic refresh, it should start its own task.

Provided Methods

fn publish(&self, _data: &EndpointData)

Publishes the given EndpointData to the EIR mechanism.

This is fire and forget, since the Endpoint can not wait for successful publishing. If publishing is async, the implementation should start it’s own task.

This will be called from a tokio task, so it is safe to spawn new tasks. These tasks will be run on the runtime of the super::Endpoint.

fn resolve( &self, _endpoint_id: EndpointId, ) -> Option<BoxStream<Result<EndpointIdResolutionItem, EndpointIdResolutionError>>>

Resolves the EndpointIdResolutionItem for the given EndpointId.

Once the returned [BoxStream] is dropped, the service should stop any pending work.