iroh

Module ers

Endpoint ID to address resolution.

To connect to an iroh endpoint a EndpointAddr is needed, which may contain a RelayUrl or one or more direct addresses in addition to the EndpointId.

Since there is a conversion from EndpointId to EndpointAddr, you can also use connect directly with a EndpointId.

For this to work however, the endpoint has to get the addressing information by other means.

Endpoint ID resolution is an automated system for an Endpoint to retrieve this addressing information. Each iroh endpoint will automatically publish their own addressing information. Usually this means publishing which RelayUrl to use for their EndpointId, but they could also publish their direct addresses.

The EndpointIdResolutionSystem trait is used to define an endpoint ID resolution system. This allows multiple implementations to co-exist because there are many possible ways to implement this. Each Endpoint can use the endpoint ID resolution mechanisms most suitable to the application. The Builder::ers method is used to add an endpoint ID resolution mechanism to an Endpoint.

Some generally useful ERS implementations are provided:

StaticProvider which allows application to add and remove out-of-band addressing information.

The ers::Dns which performs lookups via the standard DNS systems. To publish to this DNS server a PkarrPublisher is needed. Number 0 runs a public instance of a PkarrPublisher with attached DNS server which is globally available and a reliable default choice.

The PkarrResolver which can perform lookups from designated pkarr relay servers using HTTP.

ers::Mdns: mdns::MdnsEndpointIdResolution which uses the crate swarm-discovery, an opinionated mDNS implementation, to discover endpoints on the local network.

The ers::Dht also uses the pkarr system but can also publish and lookup records to/from the Mainline DHT. It requires enabling the ers-pkarr-dht feature.

To use multiple ERS’s simultaneously you can call Builder::ers. This will use ConcurrentErs under the hood, which performs lookups to all ERS systems at the same time.

Builder::ers takes any type that implements IntoErs. You can implement that trait on a builder struct if your ERS needs information from the endpoint it is mounted on. After endpoint construction, your ERS is built by calling IntoErs::into_ers, passing the finished Endpoint to your builder.

If your ERS does not need any information from its endpoint, you can pass the ERS service directly to Builder::ers: All types that implement EndpointIdResolutionSystem also have a blanket implementation of IntoErs.

Examples

A very common setup is to enable DNS ERS, which needs to be done in two parts as a PkarrPublisher and ers::Dns:

use iroh::{
    Endpoint, SecretKey,
    endpoint::RelayMode,
    ers::{self, PkarrPublisher},
};

let ep = Endpoint::empty_builder(RelayMode::Default)
    .ers(PkarrPublisher::n0_dns())
    .ers(ers::Dns::n0_dns())
    .bind()
    .await?;

To also enable ers::Mdns it can be added as another service.

#[cfg(feature = "mdns")]
let ep = Endpoint::empty_builder(RelayMode::Default)
    .ers(PkarrPublisher::n0_dns())
    .ers(ers::Dns::n0_dns())
    .ers(ers::Mdns::builder())
    .bind()
    .await?;

Re-exports

pub use dns::*;Non-wasm_browser

pub use mdns::*;mdns

pub use pkarr::dht::*;ers-pkarr-dht

pub use pkarr::*;

pub use static_provider::*;

Modules

dnsNon-wasm_browser

DNS endpoint discovery for iroh

mdnsmdns

An endpoint ID resolution service that uses an mdns-like service to discover and resolve the addresses of local endpoints.

pkarr

An endpoint ID resolution service which publishes and resolves endpoint information using a pkarr relay.

static_provider

A static endpoint id resolution system to manually add endpoint addressing information.

Structs

ConcurrentErs

An endpoint ID resolution system that combines multiple ERS sources.

EndpointData

Data about an endpoint that may be published to and resolved from discovery services.

EndpointInfo

Information about an endpoint that may be published to and resolved from discovery services.

Item

Endpoint ID resolution results from EndpointIdResolutionSystems.

UserData

Under the hood this is a UTF-8 String is no longer than UserData::MAX_LENGTH bytes.

Enums

Error

EndpointIdResolutionSystem errors

IntoErsError

IntoErs errors

ParseError

Traits

EndpointIdResolutionSystem

Endpoint ID resolution system for super::Endpoint.

IntoErs

Trait for structs that can be converted into EndpointIdResolutionSystems.