iroh::ers

Trait IntoErs

pub trait IntoErs:
    Send
    + Sync
    + Debug
    + 'static {
    // Required method
    fn into_ers(
        self,
        endpoint: &Endpoint,
    ) -> Result<impl EndpointIdResolutionSystem, IntoErsError>;
}

Trait for structs that can be converted into EndpointIdResolutionSystems.

This trait is implemented on builders for ERS’s. Any type that implements this trait can be added as a ERS in Builder::ers.

Any type that implements EndpointIdResolutionSystem also implements IntoErs.

Iroh uses this trait to allow configuring the set of ERS’s on the endpoint builder, while providing the EIR services access to information about the endpoint to IntoErs::into_ers.

Required Methods

fn into_ers( self, endpoint: &Endpoint, ) -> Result<impl EndpointIdResolutionSystem, IntoErsError>

Turns this endpoint ID resolution builder into a ready-to-use ERS.

If an error is returned, building the endpoint will fail with this error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.