Module iroh::net::discovery

Node address discovery.

To connect to an iroh-net node a NodeAddr is needed, which needs to contain either a RelayUrl or one or more direct addresses. However it is often more desirable to be able to connect with only the NodeId, as Endpoint::connect_by_node_id does.

For connecting by NodeId to work however, the endpoint has to get the addressing information by other means. This can be done by manually calling Endpoint::add_node_addr, but that still requires knowing the other addressing information.

Node discovery is an automated system for an Endpoint to retrieve this addressing information. Each iroh-net node will automatically publish their own addressing information. Usually this means publishing which RelayUrl to use for their NodeId, but they could also publish direct addresses.

The Discovery trait is used to define node discovery. This allows multiple implementations to co-exist because there are many possible ways to implement this. Each Endpoint can use the discovery mechanisms most suitable to the application. The Builder::discovery method is used to add a discovery mechanism to an Endpoint.

Some generally useful discovery implementations are provided:

The DnsDiscovery which supports publishing to a special DNS server and performs lookups via the standard DNS systems. Number 0 runs a public instance of this which is globally available and a reliable default choice.

The PkarrResolver which can perform lookups from designated pkarr relay servers using HTTP.

To use multiple discovery systems simultaneously use ConcurrentDiscovery which will perform lookups to all discovery systems at the same time.

Modules

dns

DNS node discovery for iroh-net

local_swarm_discovery

A discovery service that uses an mdns-like service to discover local nodes.

pkarr

A discovery service which publishes and resolves node information to a pkarr relay.

Structs

ConcurrentDiscovery

A discovery service that combines multiple discovery sources.

DiscoveryItem

The results returned from Discovery::resolve.

Traits

Discovery

Node discovery for super::Endpoint.