Struct iroh::net::discovery::pkarr::dht::Builder

pub struct Builder { /* private fields */ }

Builder for PkarrNodeDiscovery.

By default, publishing to the DHT is enabled, and relay publishing is disabled.

Implementations

impl Builder

pub fn client(self, client: PkarrClient) -> Builder

Explicitly set the pkarr client to use.

pub fn secret_key(self, secret_key: SecretKey) -> Builder

Set the secret key to use for signing the DNS packets.

Without a secret key, the node will not publish its address to the DHT.

pub fn ttl(self, ttl: u32) -> Builder

Set the time-to-live value for the DNS packets.

pub fn pkarr_relay(self, pkarr_relay: Url) -> Builder

Set the pkarr relay URL to use.

pub fn n0_dns_pkarr_relay(self) -> Builder

Use the default pkarr relay URL.

pub fn dht(self, dht: bool) -> Builder

Set whether to publish to the mainline DHT.

pub fn include_direct_addresses(self, include_direct_addresses: bool) -> Builder

Set whether to include the direct addresses in the DNS packet.

pub fn initial_publish_delay(self, initial_publish_delay: Duration) -> Builder

Set the initial delay before the first publish.

pub fn republish_delay(self, republish_delay: Duration) -> Builder

Set the republish delay for the DHT.

pub fn build(self) -> Result<DhtDiscovery, Error>

Build the discovery mechanism.