iroh::net_report
Function bind_local_stun_socket

pub fn bind_local_stun_socket(
    network: IpFamily,
    actor_addr: Addr,
    cancel_token: CancellationToken,
) -> Option<Arc<UdpSocket>>

Attempts to bind a local socket to send STUN packets from.

If successful this returns the bound socket and will forward STUN responses to the provided actor_addr. The cancel_token serves to stop the packet forwarding when the socket is no longer needed.