Module iroh::node

Node API

An iroh node is a server that is identified by an Ed25519 keypair and is globally reachable via the node id, which is the public key of the keypair.

By default, an iroh node speaks a number of built-in protocols. You can extend the node with custom protocols or disable built-in protocols.

Building a node

Nodes get created using the Builder which provides a very powerful API to configure every aspect of the node.

When using the default set of protocols, use spawn to spawn a node directly from the builder.

When adding custom protocols, use build to get a ProtocolBuilder that allows to add custom protocols, then call spawn to spawn the fully configured node.

To implement a custom protocol, implement the ProtocolHandler trait and use ProtocolBuilder::accept to add it to the node.

Using a node

Once created, a node offers a small number of methods to interact with it, most notably the iroh-net endpoint it is bound to.

The main way to interact with a node is through the client.

(The Node implements Deref for client, which means that methods defined on Client can be called on Node as well, without going through client)

To shut down the node, call Node::shutdown.

Structs

Builder

Builder for the Node.

Node

A server which implements the iroh node.

ProtocolBuilder

A node that is initialized but not yet spawned.

Enums

DiscoveryConfig

Configuration for node discovery.

DocsStorage

Storage backend for documents.

GcPolicy

Policy for garbage collection.

RpcStatus

The current status of the RPC endpoint.

StorageConfig

Configuration for storage.

Constants

DEFAULT_RPC_ADDR

The default bind addr of the RPC .

Traits

ProtocolHandler

Handler for incoming connections.

Type Aliases

FsNode

Persistent node.

IrohServerEndpoint

The quic-rpc server endpoint for the iroh node.

MemNode

In memory node.