Struct iroh::node::Builder

pub struct Builder<D>
where
    D: Map,
{ /* private fields */ }

Builder for the Node.

You must supply a blob store and a document store.

Blob store implementations are available in iroh_blobs::store. Document store implementations are available in iroh_docs::store.

Everything else is optional, with some sensible defaults.

The default relay servers are hosted by number 0 on the iroh.network domain. To customise this use the Builder::relay_mode function.

For node discovery the default is to use the number 0 hosted DNS server hosted on iroh.link. To customise this use the Builder::node_discovery function.

Note that some defaults change when running using cfg(test), see the individual methods for details.

Finally you can create and run the node by calling Builder::spawn.

The returned Node is awaitable to know when it finishes. It can be terminated using Node::shutdown.

Implementations

impl<D: Map> Builder<D>

pub fn with_db_and_store( blobs_store: D, docs_storage: DocsStorage, storage: StorageConfig ) -> Self

Creates a new builder for Node using the given databases.

impl<D> Builder<D>
where D: BaoStore,

pub fn blobs_events(self, blob_events: impl Into<EventSender>) -> Self

Configure a blob events sender. This will replace the previous blob event sender. By default, no events are sent.

To define an event sender, implement the iroh_blobs::provider::CustomEventSender trait.

pub async fn persist(self, root: impl AsRef<Path>) -> Result<Builder<Store>>

Persist all node data in the provided directory.

pub fn rpc_endpoint( self, value: IrohServerEndpoint, rpc_addr: Option<SocketAddr> ) -> Self

Configure rpc endpoint.

pub async fn enable_rpc(self) -> Result<Builder<D>>

Configure the default iroh rpc endpoint, on the default address.

pub async fn enable_rpc_with_addr( self, rpc_addr: SocketAddr ) -> Result<Builder<D>>

Configure the default iroh rpc endpoint.

pub fn gc_policy(self, gc_policy: GcPolicy) -> Self

Sets the garbage collection policy.

By default garbage collection is disabled.

pub fn enable_docs(self) -> Self

Enables documents support on this node.

pub fn relay_mode(self, relay_mode: RelayMode) -> Self

Sets the relay servers to assist in establishing connectivity.

Relay servers are used to discover other nodes by PublicKey and also help establish connections between peers by being an initial relay for traffic while assisting in holepunching to establish a direct connection between peers.

When using RelayMode::Custom, the provided relay_map must contain at least one configured relay node. If an invalid iroh_net::relay::RelayMode is provided Self::spawn will result in an error.

Usage during tests

Note that while the default is RelayMode::Default, when using cfg(test) or when the test-utils cargo feature RelayMode::Staging is the default.

pub fn node_discovery(self, config: DiscoveryConfig) -> Self

Sets the node discovery mechanism.

Node discovery enables connecting to other peers by only the NodeId. This usually works by the nodes publishing their RelayUrl and/or their direct addresses to some publicly available service.

See DiscoveryConfig::default for the defaults, note that the defaults change when using cfg(test).

pub fn dns_resolver(self, dns_resolver: DnsResolver) -> Self

Optionally set a custom DNS resolver to use for the magic endpoint.

The DNS resolver is used to resolve relay hostnames, and node addresses if DnsDiscovery is configured (which is the default).

By default, all magic endpoints share a DNS resolver, which is configured to use the host system’s DNS configuration. You can pass a custom instance of DnsResolver here to use a differently configured DNS resolver for this endpoint.

pub fn bind_addr_v4(self, addr: SocketAddrV4) -> Self

Binds the node service to a specific socket IPv4 address.

By default this will be set to 0.0.0.0:11204

Setting the port to 0 will assign a random port. If the port used is not free, a random different port will be used.

pub fn bind_addr_v6(self, addr: SocketAddrV6) -> Self

Binds the node service to a specific socket IPv6 address.

By default this will be set to [::]:11205

Setting the port to 0 will assign a random port. If the port used is not free, a random different port will be used.

pub fn bind_random_port(self) -> Self

Use a random port for both IPv4 and IPv6.

This is a convenience function useful when you do not need a specific port and want to avoid conflicts when running multiple instances, e.g. in tests.

This overrides the ports of the socket addresses provided by Builder::bind_addr_v4 and Builder::bind_addr_v6. By default both of those bind to the unspecified address, which would result in 0.0.0.0:11204 and [::]:11205 as bind addresses unless they are changed.

pub fn secret_key(self, secret_key: SecretKey) -> Self

Uses the given SecretKey for the PublicKey instead of a newly generated one.

pub fn transport_config(self, config: TransportConfig) -> Self

Sets a custom TransportConfig to be used by the Endpoint.

If not set, the Endpoint will use its default TransportConfig. See crate::net::endpoint::Builder::transport_config for details.

pub fn insecure_skip_relay_cert_verify(self, skip_verify: bool) -> Self

Available on test or crate feature test-utils only.

Skip verification of SSL certificates from relay servers

May only be used in tests.

pub fn register_gc_done_cb(self, cb: Box<dyn Fn() + Send>) -> Self

Available on test or crate feature test-utils only.

Register a callback for when GC is done.

pub fn keylog(self, keylog: bool) -> Self

Whether to log the SSL pre-master key.

If true and the SSLKEYLOGFILE environment variable is the path to a file this file will be used to log the SSL pre-master key. This is useful to inspect captured traffic.

pub async fn spawn(self) -> Result<Node<D>>

Spawns the Node in a tokio task.

This will create the underlying network server and spawn a tokio task accepting connections. The returned Node can be used to control the task as well as get information about it.

pub async fn build(self) -> Result<ProtocolBuilder<D>>

Builds a node without spawning it.

Returns a ProtocolBuilder, on which custom protocols can be registered with ProtocolBuilder::accept. To spawn the node, call ProtocolBuilder::spawn.