iroh::protocol

Struct Router

pub struct Router { /* private fields */ }

The built router.

Construct this using Router::builder.

When dropped, this will abort listening the tasks, so make sure to store it.

Even with this abort-on-drop behaviour, itโs recommended to call and await Router::shutdown before ending the process.

As an example for graceful shutdown, e.g. for tests or CLI tools, wait for [tokio::signal::ctrl_c()]:

let endpoint = Endpoint::bind().await?;

let router = Router::builder(endpoint)
    // .accept(&ALPN, <something>)
    .spawn();

// wait until the user wants to
tokio::signal::ctrl_c().await.context("ctrl+c")?;
router.shutdown().await.context("shutdown")?;

Implementations

impl Router

pub fn builder(endpoint: Endpoint) -> RouterBuilder

Creates a new Router using given Endpoint.

pub fn endpoint(&self) -> &Endpoint

Returns the Endpoint stored in this router.

pub fn is_shutdown(&self) -> bool

Checks if the router is already shutdown.

pub async fn shutdown(&self) -> Result<(), JoinError>

Shuts down the accept loop cleanly.

When this function returns, all ProtocolHandlers will be shutdown and Endpoint::close will have been called.

If already shutdown, it returns Ok.

If some ProtocolHandler panicked in the accept loop, this will propagate that panic into the result here.