iroh::protocol

Trait ConnectionFilter

pub trait ConnectionFilter:
    Debug
    + Send
    + Sync
    + 'static {
    // Provided methods
    fn accept_addr(
        &self,
        _addr: SocketAddr,
        _validated: bool,
    ) -> AcceptAddrOutcome { ... }
    fn accept_endpoint_id(&self, _endpoint_id: EndpointId) -> AcceptOutcome { ... }
    fn accept_alpn(
        &self,
        _endpoint_id: EndpointId,
        _alpn: &[u8],
    ) -> AcceptAlpnOutcome { ... }
}

A filter that allows rejecting incoming connections early, based on socket address, endpoint ID, or ALPN.

Provided Methods

fn accept_addr(&self, _addr: SocketAddr, _validated: bool) -> AcceptAddrOutcome

Accept an incoming connection by its socket address.

If validated is false, this is just the socket address in the UDP packet, which might be spoofed. If validated is true, the remote has responded to a retry and we know that the socket address is correct.

fn accept_endpoint_id(&self, _endpoint_id: EndpointId) -> AcceptOutcome

Accept an incoming connection by its endpoint ID.

fn accept_alpn( &self, _endpoint_id: EndpointId, _alpn: &[u8], ) -> AcceptAlpnOutcome

Accept an incoming connection by its ALPN. This can be used to implement protocol level access control without having to wrap the protocol handler.

This is called after the connection is fully established, so the filter can provide a close error code and reason.