Module iroh::rpc_protocol::blobs

Structs

AddPathRequest

A request to the node to provide the data at the given path

AddPathResponse

Wrapper around AddProgress.

AddStreamRequest

Write a blob from a byte stream

AddStreamResponse

Wrapper around AddProgress.

ConsistencyCheckRequest

A request to the node to validate the integrity of all provided data

CreateCollectionRequest

Create a collection.

CreateCollectionResponse

A response to a create collection request

DeleteRequest

Delete a blob

DownloadRequest

A request to the node to download and share the data specified by the hash.

DownloadResponse

Progress response for DownloadRequest

ExportRequest

A request to the node to download and share the data specified by the hash.

ExportResponse

Progress response for ExportRequest

ListIncompleteRequest

List all blobs, including collections

ListRequest

List all blobs, including collections

ReadAtRequest

Get the bytes for a hash

ValidateRequest

A request to the node to validate the integrity of all provided data

Enums

AddStreamUpdate

Write a blob from a byte stream

ReadAtResponse

Response to ReadAtRequest

Request

Response