Module iroh::rpc_protocol::node

Structs

AddAddrRequest

AddrRequest

ConnectionInfoRequest

Get connection information about a specific node

ConnectionInfoResponse

A response to a connection request

ConnectionsRequest

List connection information about all the nodes we know about

ConnectionsResponse

A response to a connections request

CounterStats

Counter stats

IdRequest

A request to get information the identity of the node.

NodeWatchRequest

A request to watch for the node status

RelayRequest

ShutdownRequest

A request to shutdown the node

StatsRequest

Get stats for the running Iroh node

StatsResponse

Response to StatsRequest

StatusRequest

A request to get information about the status of the node.

VersionResponse

The response to a version request

WatchResponse

The response to a watch request

Enums

Request

Response