Trait iroh::spaces::store::traits::EntryStorage

pub trait EntryStorage: EntryReader + Clone + Debug + 'static {
    type Reader: EntryReader;
    type Snapshot: EntryReader + Clone;

    // Required methods
    fn reader(&self) -> Self::Reader;
    fn snapshot(&self) -> Result<Self::Snapshot, Error>;
    fn ingest_entry(
        &self,
        entry: &AuthorisedEntry<iroh_willow::::proto::data_model::AuthorisedEntry::{constant#0}, iroh_willow::::proto::data_model::AuthorisedEntry::{constant#1}, iroh_willow::::proto::data_model::AuthorisedEntry::{constant#2}, NamespaceId, UserId, PayloadDigest, McAuthorisationToken<iroh_willow::::proto::meadowcap::McAuthorisationToken::{constant#0}, iroh_willow::::proto::meadowcap::McAuthorisationToken::{constant#1}, iroh_willow::::proto::meadowcap::McAuthorisationToken::{constant#2}, NamespaceId, NamespaceSignature, UserId, UserSignature>>,
        origin: EntryOrigin
    ) -> Result<bool, Error>;
    fn subscribe_area(
        &self,
        namespace: NamespaceId,
        area: Area<iroh_willow::::proto::grouping::Area::{constant#0}, iroh_willow::::proto::grouping::Area::{constant#1}, iroh_willow::::proto::grouping::Area::{constant#2}, UserId>,
        params: SubscribeParams
    ) -> impl Stream<Item = StoreEvent> + Unpin + 'static;
    fn resume_subscription(
        &self,
        progress_id: u64,
        namespace: NamespaceId,
        area: Area<iroh_willow::::proto::grouping::Area::{constant#0}, iroh_willow::::proto::grouping::Area::{constant#1}, iroh_willow::::proto::grouping::Area::{constant#2}, UserId>,
        params: SubscribeParams
    ) -> impl Stream<Item = StoreEvent> + Unpin + 'static;
}

Storage for entries.

Required Associated Types

type Reader: EntryReader

type Snapshot: EntryReader + Clone

Required Methods

fn reader(&self) -> Self::Reader

fn snapshot(&self) -> Result<Self::Snapshot, Error>

fn ingest_entry( &self, entry: &AuthorisedEntry<iroh_willow::::proto::data_model::AuthorisedEntry::{constant#0}, iroh_willow::::proto::data_model::AuthorisedEntry::{constant#1}, iroh_willow::::proto::data_model::AuthorisedEntry::{constant#2}, NamespaceId, UserId, PayloadDigest, McAuthorisationToken<iroh_willow::::proto::meadowcap::McAuthorisationToken::{constant#0}, iroh_willow::::proto::meadowcap::McAuthorisationToken::{constant#1}, iroh_willow::::proto::meadowcap::McAuthorisationToken::{constant#2}, NamespaceId, NamespaceSignature, UserId, UserSignature>>, origin: EntryOrigin ) -> Result<bool, Error>

Ingest a new entry.

Returns true if the entry was ingested, and false if the entry was not ingested because a newer entry exists.

fn subscribe_area( &self, namespace: NamespaceId, area: Area<iroh_willow::::proto::grouping::Area::{constant#0}, iroh_willow::::proto::grouping::Area::{constant#1}, iroh_willow::::proto::grouping::Area::{constant#2}, UserId>, params: SubscribeParams ) -> impl Stream<Item = StoreEvent> + Unpin + 'static

Subscribe to events concerning entries included by an [AreaOfInterest], returning a producer of StoreEvents which occurred since the moment of calling this function.

If ignore_incomplete_payloads is true, the producer will not produce entries with incomplete corresponding payloads. If ignore_empty_payloads is true, the producer will not produce entries with a payload_length of 0.

fn resume_subscription( &self, progress_id: u64, namespace: NamespaceId, area: Area<iroh_willow::::proto::grouping::Area::{constant#0}, iroh_willow::::proto::grouping::Area::{constant#1}, iroh_willow::::proto::grouping::Area::{constant#2}, UserId>, params: SubscribeParams ) -> impl Stream<Item = StoreEvent> + Unpin + 'static

Attempt to resume a subscription using a progress ID obtained from a previous subscription, or return an error if this store implementation is unable to resume the subscription.

Object Safety

This trait is not object safe.