iroh

Struct EndpointAddr

pub struct EndpointAddr {
    pub id: EndpointId,
    pub addrs: BTreeSet<TransportAddr>,
}

Network-level addressing information for an iroh endpoint.

This combines an endpoint’s identifier with network-level addressing information of how to contact the endpoint.

To establish a network connection to an endpoint both the EndpointId and one or more network paths are needed. The network paths can come from various sources, current sources can come from

A discovery service which can provide routing information for a given EndpointId.

A RelayUrl of the endpoint’s home relay, this allows establishing the connection via the Relay server and is very reliable.

One or more IP based addresses on which the endpoint might be reachable. Depending on the network location of both endpoints it might not be possible to establish a direct connection without the help of a Relay server.

This structure will always contain the required EndpointId and will contain an optional number of other addressing information. It is a generic addressing type used whenever a connection to other endpoints needs to be established.

Fields
id: EndpointId

The endpoint’s identifier.

addrs: BTreeSet<TransportAddr>

The endpoint’s addresses

Implementations

impl EndpointAddr

pub fn new(id: impl Into<EndpointId>) -> EndpointAddr

Creates a new EndpointAddr with no network level addresses.

This still is usable with e.g. a discovery service to establish a connection, depending on the situation.

pub fn from_parts( id: impl Into<EndpointId>, addrs: impl IntoIterator<Item = TransportAddr>, ) -> EndpointAddr

Creates a new EndpointAddr from its parts.

pub fn with_relay_url(self, relay_url: RelayUrl) -> EndpointAddr

Adds a RelayUrl address.

pub fn with_ip_addr(self, addr: SocketAddr) -> EndpointAddr

Adds an IP based address.

pub fn with_addrs( self, addrs: impl IntoIterator<Item = TransportAddr>, ) -> EndpointAddr

Adds a list of addresses.

pub fn is_empty(&self) -> bool

Returns true, if only a EndpointId is present.

pub fn ip_addrs(&self) -> impl Iterator<Item = &SocketAddr>

Returns a list of IP addresses of this peer.

pub fn relay_urls(&self) -> impl Iterator<Item = &RelayUrl>

Returns a list of relay urls of this peer.

In practice this is expected to be zero or one home relay for all known cases currently.