iroh

Struct NodeAddr

pub struct NodeAddr {
    pub node_id: PublicKey,
    pub relay_url: Option<RelayUrl>,
    pub direct_addresses: BTreeSet<SocketAddr>,
}

Network-level addressing information for an iroh node.

This combines a node’s identifier with network-level addressing information of how to contact the node.

To establish a network connection to a node both the NodeId and one or more network paths are needed. The network paths can come from various sources:

A discovery service which can provide routing information for a given NodeId.

A RelayUrl of the node’s home relay, this allows establishing the connection via the Relay server and is very reliable.

One or more direct addresses on which the node might be reachable. Depending on the network location of both nodes it might not be possible to establish a direct connection without the help of a Relay server.

This structure will always contain the required NodeId and will contain an optional number of network-level addressing information. It is a generic addressing type used whenever a connection to other nodes needs to be established.

Fields
node_id: PublicKey

The node’s identifier.

relay_url: Option<RelayUrl>

The node’s home relay url.

direct_addresses: BTreeSet<SocketAddr>

Socket addresses where the peer might be reached directly.

Implementations

impl NodeAddr

pub fn new(node_id: PublicKey) -> NodeAddr

Creates a new NodeAddr with no relay_url and no direct_addresses.

pub fn with_relay_url(self, relay_url: RelayUrl) -> NodeAddr

Adds a relay url.

pub fn with_direct_addresses( self, addresses: impl IntoIterator<Item = SocketAddr>, ) -> NodeAddr

Adds the given direct addresses.

pub fn from_parts( node_id: PublicKey, relay_url: Option<RelayUrl>, direct_addresses: impl IntoIterator<Item = SocketAddr>, ) -> NodeAddr

Creates a new NodeAddr from its parts.

pub fn is_empty(&self) -> bool

Returns true, if only a NodeId is present.

pub fn direct_addresses(&self) -> impl Iterator<Item = &SocketAddr>

Returns the direct addresses of this peer.

pub fn relay_url(&self) -> Option<&RelayUrl>

Returns the relay url of this peer.