iroh

Struct RelayUrl

pub struct RelayUrl(/* private fields */);

A URL identifying a relay server.

It is cheaply clonable, as the underlying type is wrapped into an Arc. The main type under the hood though is Url, with a few custom tweaks:

A relay URL is never a relative URL, so an implicit . is added at the end of the domain name if missing.

fmt::Debug is implemented so it prints the URL rather than the URL struct fields. Useful when logging e.g. Option<RelayUrl>.

To create a RelayUrl use the From<Url> implementation.