iroh::test_utils
Function run_relay_server_with

pub async fn run_relay_server_with(
    stun: Option<StunConfig>,
    quic: bool,
) -> Result<(RelayMap, RelayUrl, Server)>

Available on crate feature test-utils only.

Runs a relay server.

stun can be set to None to disable stun, or set to Some StunConfig, to enable stun on a specific socket.

If quic is set to true, it will make the appropriate QuicConfig from the generated tls certificates and run the quic server at a random free port.

The return value is similar to run_relay_server.