iroh::test_utils

Struct DnsPkarrServer

pub struct DnsPkarrServer {
    pub node_origin: String,
    pub nameserver: SocketAddr,
    pub pkarr_url: Url,
    /* private fields */
}

Available on crate feature test-utils only.

Handle and drop guard for test DNS and Pkarr servers.

Once the struct is dropped the servers will shut down.

Fields
node_origin: String

The node origin domain.

nameserver: SocketAddr

The socket address of the DNS server.

pkarr_url: Url

The HTTP URL of the Pkarr server.

Implementations

impl DnsPkarrServer

pub async fn run() -> Result<Self>

Run DNS and Pkarr servers on localhost.

pub async fn run_with_origin(node_origin: String) -> Result<Self>

Run DNS and Pkarr servers on localhost with the specified node_origin domain.

pub fn discovery(&self, secret_key: SecretKey) -> ConcurrentDiscovery

Create a ConcurrentDiscovery with DnsDiscovery and PkarrPublisher configured to use the test servers.

pub fn dns_resolver(&self) -> DnsResolver

Create a DnsResolver configured to use the test DNS server.

pub async fn on_node(&self, node_id: &NodeId, timeout: Duration) -> Result<()>

Wait until a Pkarr announce for a node is published to the server.

If timeout elapses an error is returned.