Module iroh::tls::certificate

X.509 certificate handling.

This module handles generation, signing, and verification of certificates.

Based on rust-libp2p/transports/tls/src/certificate.rs originally licensed under MIT by Parity Technologies (UK) Ltd.

Structs

GenError

An error that occurs during certificate generation.

P2pCertificate

An X.509 certificate with a libp2p-specific extension is used to secure libp2p connections.

P2pExtension

The contents of the specific libp2p extension, containing the public host key and a signature performed using the private host key.

ParseError

An error that occurs during certificate parsing.

VerificationError

An error that occurs during signature verification.

Functions

generate

Generates a self-signed TLS certificate that includes a libp2p-specific certificate extension containing the public key of the given secret key.

parse

Attempts to parse the provided bytes as a P2pCertificate.