iroh

Type Alias EndpointId

pub type EndpointId = PublicKey;

The identifier for an endpoint in the (iroh) network.

Each endpoint in iroh has a unique identifier created as a cryptographic key. This can be used to globally identify an endpoint. Since it is also a cryptographic key it is also the mechanism by which all traffic is always encrypted for a specific endpoint only.

This is equivalent to PublicKey. By convention we will (or should) use PublicKey as type name when performing cryptographic operations, but use EndpointId when referencing an endpoint. E.g.:

encrypt(key: PublicKey)

send_to(endpoint: EndpointId)

Aliased Type

pub struct EndpointId(/* private fields */);