Module iroh::util::fs

Utilities for filesystem operations.

Structs

PathContent

Information about the content on a path

Functions

canonicalized_path_to_string

This function converts an already canonicalized path to a string.

key_to_path

Helper function that translates a key that was derived from the path_to_key function back into a path.

load_secret_key

Loads a SecretKey from the provided file, or stores a newly generated one at the given location.

path_content_info

Walks the directory to get the total size and number of files in directory or file

path_to_key

Helper function that creates a document key from a canonicalized path, removing the root and adding the prefix, if they exist