Module iroh::util::progress

Generic utilities to track progress of data transfers.

This is not especially specific to iroh but can be helpful together with it. The ProgressEmitter has a ProgressEmitter::wrap_async_read method which can make it easy to track process of transfers.

However based on your environment there might also be better choices for this, e.g. very similar and more advanced functionality is available in the indicatif crate for terminal applications.

Structs

Progress

A sender for progress messages.

ProgressAsyncReader

A wrapper around [AsyncRead] which increments a ProgressEmitter.

ProgressEmitter

A generic progress event emitter.

ProgressWriter

A writer that tries to send the total number of bytes written after each write