Module iroh::util::watchable

Watchable values.

A Watchable exists to keep track of a value which may change over time. It allows observers to be notified of changes to the value. The aim is to always be aware of the last value, not to observe every value there has ever been.

Structs

Disconnected

The error for when a Watcher is disconnected from its underlying Watchable value, because that value was dropped.

WatchNextFut

Future returning another item from a Watcher.

Watchable

A value who’s changes over time can be observed.

Watcher

An observer for a value.

WatcherStream

A stream for a Watcher’s next values.