iroh
Module watchable

Watchable values.

A Watchable exists to keep track of a value which may change over time. It allows observers to be notified of changes to the value. The aim is to always be aware of the last value, not to observe every value change.

In that way, a Watchable is like a [tokio::sync::broadcast::Sender] (and a Watcher is like a [tokio::sync::broadcast::Receiver]), except that there’s no risk of the channel filling up, but instead you might miss items.

Structs

Disconnected

The error for when a Watcher is disconnected from its underlying Watchable value, because of that watchable having been dropped.

WatchInitializedFut

Future returning the current or next value that’s Some value. in a Watcher.

WatchNextFut

Future returning the next item after the current one in a Watcher.

Watchable

A value who’s changes over time can be observed.

Watcher

An observer for a value.

WatcherStream

A stream for a Watcher’s next values.