iroh::watchable
Trait Watcher

pub trait Watcher: Clone {
    type Value: Clone + Eq;

    // Required methods
    fn get(&self) -> Result<Self::Value, Disconnected>;
    fn poll_updated(
        &mut self,
        cx: &mut Context<'_>,
    ) -> Poll<Result<Self::Value, Disconnected>>;

    // Provided methods
    fn updated(&mut self) -> WatchNextFut<'_, Self>  { ... }
    fn initialized<T>(&mut self) -> WatchInitializedFut<'_, T, Self> 
       where Self: Watcher<Value = Option<T>> { ... }
    fn stream(self) -> WatcherStream<Self>
       where Self: Unpin { ... }
    fn stream_updates_only(self) -> WatcherStream<Self>
       where Self: Unpin { ... }
    fn map<T: Clone + Eq, F: Fn(Self::Value) -> T>(
        self,
        map: F,
    ) -> Result<MapWatcher<Self, T, F>, Disconnected> { ... }
    fn or<W: Watcher>(self, other: W) -> OrWatcher<Self, W> { ... }
}

A handle to a value that’s represented by one or more underlying Watchables.

This handle allows one to observe the latest state and be notified of any changes to this value.

Required Associated Types

type Value: Clone + Eq

The type of value that can change.

We require Clone, because we need to be able to make the values have a lifetime that’s detached from the original Watchable’s lifetime.

We require Eq, to be able to check whether the value actually changed or not, so we can notify or not notify accordingly.

Required Methods

fn get(&self) -> Result<Self::Value, Disconnected>

Returns the current state of the underlying value, or errors out with Disconnected, if one of the underlying Watchables has been dropped.

fn poll_updated( &mut self, cx: &mut Context<'_>, ) -> Poll<Result<Self::Value, Disconnected>>

Polls for the next value, or returns Disconnected if one of the underlying watchables has been dropped.

Provided Methods

fn updated(&mut self) -> WatchNextFut<'_, Self>

Returns a future completing with Ok(value) once a new value is set, or with Err(Disconnected) if the connected Watchable was dropped.

Cancel Safety

The returned future is cancel-safe.

fn initialized<T>(&mut self) -> WatchInitializedFut<'_, T, Self>
where Self: Watcher<Value = Option<T>>,

Returns a future completing once the value is set to Some value.

If the current value is Some value, this future will resolve immediately.

This is a utility for the common case of storing an Option inside a Watchable.

fn stream(self) -> WatcherStream<Self>
where Self: Unpin,

Returns a stream which will yield the most recent values as items.

The first item of the stream is the current value, so that this stream can be easily used to operate on the most recent value.

Note however, that only the last item is stored. If the stream is not polled when an item is available it can be replaced with another item by the time it is polled.

This stream ends once the original Watchable has been dropped.

Cancel Safety

The returned stream is cancel-safe.

fn stream_updates_only(self) -> WatcherStream<Self>
where Self: Unpin,

Returns a stream which will yield the most recent values as items, starting from the next unobserved future value.

This means this stream will only yield values when the watched value changes, the value stored at the time the stream is created is not yielded.

Note however, that only the last item is stored. If the stream is not polled when an item is available it can be replaced with another item by the time it is polled.

This stream ends once the original Watchable has been dropped.

Cancel Safety

The returned stream is cancel-safe.

fn map<T: Clone + Eq, F: Fn(Self::Value) -> T>( self, map: F, ) -> Result<MapWatcher<Self, T, F>, Disconnected>

Maps this watcher with a function that transforms the observed values.

fn or<W: Watcher>(self, other: W) -> OrWatcher<Self, W>

Returns a watcher that updates every time this or the other watcher updates, and yields both watcher’s items together when that happens.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.