iroh
Module defaults

Default values used in iroh

Modules

prod

Production configuration.

staging

Staging configuration.

Constants

DEFAULT_HTTPS_PORT

The default HTTPS port used by the Relay server.

DEFAULT_HTTP_PORT

The default HTTP port used by the Relay server.

DEFAULT_METRICS_PORT

The default metrics port used by the Relay server.

DEFAULT_RELAY_QUIC_PORT

The default QUIC port used by the Relay server to accept QUIC connections for QUIC address discovery

DEFAULT_STUN_PORT

The default STUN port used by the Relay server.