iroh::discovery

Module mdns

Available on crate feature discovery-local-network only.

A discovery service that uses an mdns-like service to discover local nodes.

This allows you to use an mdns-like swarm discovery service to find address information about nodes that are on your local network, no relay or outside internet needed. See the swarm-discovery crate for more details.

When MdnsDiscovery is enabled, it’s possible to get a list of the locally discovered nodes by filtering a list of RemoteInfos.

use std::time::Duration;

use iroh::{
    SecretKey,
    discovery::mdns::{DiscoveryEvent, MdnsDiscovery},
    endpoint::{Endpoint, Source},
};
use n0_future::StreamExt;

#[tokio::main]
async fn main() {
    let recent = Duration::from_secs(600); // 10 minutes in seconds
    let endpoint = Endpoint::builder().bind().await.unwrap();

    // Register the discovery services with the endpoint
    let mdns = MdnsDiscovery::builder().build(endpoint.node_id()).unwrap();
    endpoint.discovery().add(mdns.clone());

    // Subscribe to the discovery events
    let mut events = mdns.subscribe().await;
    while let Some(event) = events.next().await {
        match event {
            DiscoveryEvent::Discovered { node_info, .. } => {
                println!("MDNS discovered: {:?}", node_info);
            }
            DiscoveryEvent::Expired { node_id } => {
                println!("MDNS expired: {node_id}");
            }
        }
    }
}

Structs

MdnsDiscovery

Discovery using swarm-discovery, a variation on mdns

MdnsDiscoveryBuilder

Builder for MdnsDiscovery.

Enums

DiscoveryEvent

An event emitted from the MdnsDiscovery service.

Constants

NAME

Name of this discovery service.