iroh::discovery

Trait IntoDiscovery

pub trait IntoDiscovery:
    Send
    + Sync
    + Debug
    + 'static {
    // Required method
    fn into_discovery(
        self,
        endpoint: &Endpoint,
    ) -> Result<impl Discovery, IntoDiscoveryError>;
}

Trait for structs that can be converted into Discovery.

This trait is implemented on builders for discovery services. Any type that implements this trait can be added as a discovery service in Builder::discovery.

Any type that implements Discovery also implements IntoDiscovery.

Iroh uses this trait to allow configuring the set of discovery services on the endpoint builder, while providing the discovery services access to information about the endpoint to IntoDiscovery::into_discovery.

Required Methods

fn into_discovery( self, endpoint: &Endpoint, ) -> Result<impl Discovery, IntoDiscoveryError>

Turns this discovery builder into a ready-to-use discovery service.

If an error is returned, building the endpoint will fail with this error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.