iroh

Module dns

Available on non-wasm_browser only.

This module exports a DNS resolver, which is also the default resolver used in the crate::Endpoint if no custom resolver is configured.

The resolver provides methods to resolve domain names to ipv4 and ipv6 addresses, and to resolve node ids to node addresses.

See the node_info module documentation for details on how iroh node records are structured.

Structs

DnsResolver

The DNS resolver used throughout iroh.

Constants

N0_DNS_NODE_ORIGIN_PROD

The n0 testing DNS node origin, for production.

N0_DNS_NODE_ORIGIN_STAGING

The n0 testing DNS node origin, for testing.