iroh::endpoint

Struct VarInt

pub struct VarInt(/* private fields */);

An integer less than 2^62

Values of this type are suitable for encoding as QUIC variable-length integer.

Implementations

impl VarInt

pub const MAX: VarInt

The largest representable value

pub const MAX_SIZE: usize = 8usize

The largest encoded value length

pub const fn from_u32(x: u32) -> VarInt

Construct a VarInt infallibly

pub fn from_u64(x: u64) -> Result<VarInt, VarIntBoundsExceeded>

Succeeds iff x < 2^62

pub const unsafe fn from_u64_unchecked(x: u64) -> VarInt

Create a VarInt without ensuring it’s in range

Safety

x must be less than 2^62.

pub const fn into_inner(self) -> u64

Extract the integer value

pub fn saturating_add(self, rhs: impl Into<VarInt>) -> VarInt

Saturating integer addition. Computes self + rhs, saturating at the numeric bounds instead of overflowing.