iroh::metrics

Struct MagicsockMetrics

#[non_exhaustive]
pub struct MagicsockMetrics {

    pub update_direct_addrs: Counter,
    pub send_ipv4: Counter,
    pub send_ipv6: Counter,
    pub send_relay: Counter,
    pub send_relay_error: Counter,
    pub send_data: Counter,
    pub send_data_network_down: Counter,
    pub recv_data_relay: Counter,
    pub recv_data_ipv4: Counter,
    pub recv_data_ipv6: Counter,
    pub recv_datagrams: Counter,
    pub recv_gro_datagrams: Counter,
    pub send_disco_udp: Counter,
    pub send_disco_relay: Counter,
    pub sent_disco_udp: Counter,
    pub sent_disco_relay: Counter,
    pub sent_disco_ping: Counter,
    pub sent_disco_pong: Counter,
    pub sent_disco_call_me_maybe: Counter,
    pub recv_disco_bad_key: Counter,
    pub recv_disco_bad_parse: Counter,
    pub recv_disco_udp: Counter,
    pub recv_disco_relay: Counter,
    pub recv_disco_ping: Counter,
    pub recv_disco_pong: Counter,
    pub recv_disco_call_me_maybe: Counter,
    pub recv_disco_call_me_maybe_bad_disco: Counter,
    pub relay_home_change: Counter,
    pub num_direct_conns_added: Counter,
    pub num_direct_conns_removed: Counter,
    pub num_relay_conns_added: Counter,
    pub num_relay_conns_removed: Counter,
    pub actor_tick_main: Counter,
    pub actor_tick_msg: Counter,
    pub actor_tick_re_stun: Counter,
    pub actor_tick_portmap_changed: Counter,
    pub actor_tick_direct_addr_heartbeat: Counter,
    pub actor_link_change: Counter,
    pub actor_tick_other: Counter,
    pub nodes_contacted: Counter,
    pub nodes_contacted_directly: Counter,
    pub connection_handshake_success: Counter,
    pub connection_became_direct: Counter,

}

Enum of metrics for the module

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

update_direct_addrs: Countersend_ipv4: Countersend_ipv6: Countersend_relay: Countersend_relay_error: Countersend_data: Countersend_data_network_down: Counterrecv_data_relay: Counterrecv_data_ipv4: Counterrecv_data_ipv6: Counterrecv_datagrams: Counter

Number of QUIC datagrams received.

recv_gro_datagrams: Counter

Number of datagrams received using GRO

send_disco_udp: Countersend_disco_relay: Countersent_disco_udp: Countersent_disco_relay: Countersent_disco_ping: Countersent_disco_pong: Countersent_disco_call_me_maybe: Counterrecv_disco_bad_key: Counterrecv_disco_bad_parse: Counterrecv_disco_udp: Counterrecv_disco_relay: Counterrecv_disco_ping: Counterrecv_disco_pong: Counterrecv_disco_call_me_maybe: Counterrecv_disco_call_me_maybe_bad_disco: Counterrelay_home_change: Counternum_direct_conns_added: Counter

The number of direct connections we have made to peers.

num_direct_conns_removed: Counter

The number of direct connections we have lost to peers.

num_relay_conns_added: Counter

The number of connections to peers we have added over relay.

num_relay_conns_removed: Counter

The number of connections to peers we have removed over relay.

actor_tick_main: Counteractor_tick_msg: Counteractor_tick_re_stun: Counteractor_tick_portmap_changed: Counteractor_tick_direct_addr_heartbeat: Counteractor_link_change: Counteractor_tick_other: Counternodes_contacted: Counter

Number of nodes we have attempted to contact.

nodes_contacted_directly: Counter

Number of nodes we have managed to contact directly.

connection_handshake_success: Counter

Number of connections with a successful handshake.

connection_became_direct: Counter

Number of connections with a successful handshake that became direct.