iroh
Module net_report

Checks the network conditions from the current host.

NetReport is responsible for finding out the network conditions of the current host, like whether it is connected to the internet via IPv4 and/or IPv6, what the NAT situation is etc and reachability to the configured relays.

Structs

Addr

Sender to the main service.

Client

Client to run net_reports.

Metrics

Enum of metrics for the module

OptionsNon-wasm_browser

Options for running probes

QuicConfig

Pieces needed to do QUIC address discovery.

RelayLatencies

Latencies per relay node.

Report

A net_report report.

Functions

bind_local_stun_socketNon-wasm_browser

Attempts to bind a local socket to send STUN packets from.