iroh

Module net_report

Checks the network conditions from the current host.

NetReport is responsible for finding out the network conditions of the current host, like whether it is connected to the internet via IPv4 and/or IPv6, what the NAT situation is etc and reachability to the configured relays.

Structs

Metrics

Enum of metrics for the module

Options

Options for running probes

QuicConfig

Pieces needed to do QUIC address discovery.

RelayLatencies

Latencies per relay endpoint.

Report

A net_report report.

Enums

Probe

The protocol used to time an endpoint’s latency.

Constants

TIMEOUT

The maximum amount of time, in seconds, the net_report will spend gathering a single report.