Trait iroh::protocol::ProtocolHandler

pub trait ProtocolHandler: Send + Sync + Debug + 'static {
    // Required method
    fn accept(&self, conn: Connecting) -> BoxedFuture<Result<()>>;

    // Provided method
    fn shutdown(&self) -> BoxedFuture<()> { ... }
}

Handler for incoming connections.

A router accepts connections for arbitrary ALPN protocols.

With this trait, you can handle incoming connections for any protocol.

Implement this trait on a struct that should handle incoming connections. The protocol handler must then be registered on the node for an ALPN protocol with crate::protocol::RouterBuilder::accept.

Required Methods

fn accept(&self, conn: Connecting) -> BoxedFuture<Result<()>>

Handle an incoming connection.

This runs on a freshly spawned tokio task so this can be long-running.

Provided Methods

fn shutdown(&self) -> BoxedFuture<()>

Called when the node shuts down.