iroh::protocol
Trait ProtocolHandler

pub trait ProtocolHandler:
    Send
    + Sync
    + Debug
    + 'static {
    // Required method
    fn accept(&self, connection: Connection) -> BoxFuture<Result<()>>;

    // Provided methods
    fn on_connecting(
        &self,
        connecting: Connecting,
    ) -> BoxFuture<Result<Connection>> { ... }
    fn shutdown(&self) -> BoxFuture<()> { ... }
}

Handler for incoming connections.

A router accepts connections for arbitrary ALPN protocols.

With this trait, you can handle incoming connections for any protocol.

Implement this trait on a struct that should handle incoming connections. The protocol handler must then be registered on the node for an ALPN protocol with crate::protocol::RouterBuilder::accept.

Required Methods

fn accept(&self, connection: Connection) -> BoxFuture<Result<()>>

Handle an incoming connection.

This runs on a freshly spawned tokio task so this can be long-running.

Provided Methods

fn on_connecting(&self, connecting: Connecting) -> BoxFuture<Result<Connection>>

Optional interception point to handle the Connecting state.

fn shutdown(&self) -> BoxFuture<()>

Called when the node shuts down.