iroh::test_utils

Function run_relay_server_with

pub async fn run_relay_server_with(
    quic: bool,
) -> Result<(RelayMap, RelayUrl, Server), SpawnError>

Available on crate features test-utils only.

Runs a relay server.

If quic is set to true, it will make the appropriate QuicConfig from the generated tls certificates and run the quic server at a random free port.

The return value is similar to run_relay_server.